Modularize risk logic into sub-app
