+++ mainappsrc/AutoML.py
@@ -387,7 +387,6 @@ import builtins
 try:  # pragma: no cover - support direct module import
     from .user_manager import UserManager
     from .project_manager import ProjectManager
-    from .diagram_controller import DiagramController
     from .sotif_manager import SOTIFManager
 except Exception:  # pragma: no cover
     import os, sys
@@ -396,17 +395,20 @@ except Exception:  # pragma: no cover
     sys.path.append(os.path.dirname(base))
     from user_manager import UserManager
     from project_manager import ProjectManager
-    from diagram_controller import DiagramController
     from sotif_manager import SOTIFManager
 from user_manager import UserManager
 from project_manager import ProjectManager
 from cyber_manager import CyberSecurityManager
-from diagram_controller import DiagramController
 from cta_manager import ControlTreeManager
 from config.automl_constants import (
     dynamic_recommendations,
     WORK_PRODUCT_INFO as BASE_WORK_PRODUCT_INFO,
     WORK_PRODUCT_PARENTS as BASE_WORK_PRODUCT_PARENTS,
+    PMHF_TARGETS,
+    VALID_SUBTYPES,
+    AUTHOR,
+    AUTHOR_EMAIL,
+    AUTHOR_LINKEDIN,
 )
 
 builtins.REQUIREMENT_WORK_PRODUCTS = REQUIREMENT_WORK_PRODUCTS
@@ -465,6 +467,34 @@ from analysis.utils import (
 )
 from analysis.safety_management import SafetyManagementToolbox, ACTIVE_TOOLBOX
 from analysis.causal_bayesian_network import CausalBayesianNetwork, CausalBayesianNetworkDoc
+try:  # pragma: no cover - support direct module import
+    from .style_subapp import StyleSubApp
+    from .tree_subapp import TreeSubApp
+    from .diagram_export_subapp import DiagramExportSubApp
+    from .requirements_manager import RequirementsManagerSubApp
+    from .use_case_diagram_subapp import UseCaseDiagramSubApp
+    from .activity_diagram_subapp import ActivityDiagramSubApp
+    from .block_diagram_subapp import BlockDiagramSubApp
+    from .internal_block_diagram_subapp import InternalBlockDiagramSubApp
+    from .control_flow_diagram_subapp import ControlFlowDiagramSubApp
+    from .fta_subapp import FTASubApp
+    from .risk_assessment_subapp import RiskAssessmentSubApp
+    from .reliability_subapp import ReliabilitySubApp
+    from .version import VERSION
+except Exception:  # pragma: no cover
+    from style_subapp import StyleSubApp
+    from tree_subapp import TreeSubApp
+    from diagram_export_subapp import DiagramExportSubApp
+    from requirements_manager import RequirementsManagerSubApp
+    from use_case_diagram_subapp import UseCaseDiagramSubApp
+    from activity_diagram_subapp import ActivityDiagramSubApp
+    from block_diagram_subapp import BlockDiagramSubApp
+    from internal_block_diagram_subapp import InternalBlockDiagramSubApp
+    from control_flow_diagram_subapp import ControlFlowDiagramSubApp
+    from fta_subapp import FTASubApp
+    from risk_assessment_subapp import RiskAssessmentSubApp
+    from reliability_subapp import ReliabilitySubApp
+    from version import VERSION
 try:  # pragma: no cover
     from .models.fault_tree_node import FaultTreeNode
 except Exception:  # pragma: no cover
@@ -475,18 +505,10 @@ except Exception:  # pragma: no cover
     from models.fault_tree_node import FaultTreeNode
 
 from gui.toolboxes import (
-    ReliabilityWindow,
-    FI2TCWindow,
-    HazopWindow,
-    RiskAssessmentWindow,
-    TC2FIWindow,
-    HazardExplorerWindow,
     RequirementsExplorerWindow,
     DiagramElementDialog,
     _RequirementRelationDialog,
 )
-from gui.stpa_window import StpaWindow
-from gui.threat_window import ThreatWindow
 
 
 def format_requirement(req, include_id=True):
@@ -508,54 +530,7 @@ def format_requirement(req, include_id=True):
 
 from pathlib import Path
 
-
-def get_version() -> str:
-    """Read the tool version from the first line of README.md.
-
-    The README is located alongside this file so we resolve the path relative
-    to ``__file__``.  This avoids returning ``"Unknown"`` when the current
-    working directory is different (e.g. when launching from another folder or
-    from an installed package).
-    """
-    try:
-        readme = Path(__file__).resolve().parent / "README.md"
-        with readme.open("r", encoding="utf-8") as f:
-            first_line = f.readline().strip()
-            if first_line.lower().startswith("version:"):
-                return first_line.split(":", 1)[1].strip()
-    except Exception:
-        pass
-    return "Unknown"
-
-
-VERSION = get_version()
-
-# Contact information for splash screen
-AUTHOR = "Miguel Marina"
-AUTHOR_EMAIL = "karel.capek.robotics@gmail.com"
-AUTHOR_LINKEDIN = "https://www.linkedin.com/in/progman32/"
-
 from gui.dialogs.user_info_dialog import UserInfoDialog
-
-# Target PMHF limits per ASIL level (events per hour)
-PMHF_TARGETS = {
-    "D": 1e-8,
-    "C": 1e-7,
-    "B": 1e-7,
-    "A": 1e-6,
-    "QM": 1.0,
-}
-
-##########################################
-# VALID_SUBTYPES dictionary
-##########################################
-VALID_SUBTYPES = {
-    "Confidence": ["Function", "Human Task"],
-    "Robustness": ["Function", "Human Task"],
-    "Maturity": ["Functionality"],
-    "Rigor": ["Failure", "AI Error", "Functional Insufficiency"],
-    "Prototype Assurance Level (PAL)": ["Vehicle Level Function"]
-}
 
 # Node types treated as gates when rendering and editing
 _CONFIG_PATH = Path(__file__).resolve().parent / "config/diagram_rules.json"
@@ -662,197 +637,15 @@ class AutoMLApp:
         self.rc_dragged = False
         self.diagram_font = tkFont.Font(family="Arial", size=int(8 * self.zoom))
         self.style = ttk.Style()
-        try:
-            self.style.theme_use("clam")
-        except tk.TclError:
-            pass
-        self.style.configure(
-            "Treeview",
-            font=("Arial", 10),
-            background="#ffffff",
-            fieldbackground="#ffffff",
-            foreground="black",
-            borderwidth=1,
-            relief="sunken",
-        )
-        self.style.configure(
-            "Treeview.Heading",
-            background="#b5bdc9",
-            foreground="black",
-            relief="raised",
-        )
-        self.style.map(
-            "Treeview.Heading",
-            background=[("active", "#4a6ea9"), ("!active", "#b5bdc9")],
-            foreground=[("active", "white"), ("!active", "black")],
-        )
-        # ------------------------------------------------------------------
-        # Global color theme inspired by Windows classic / Windows 7
-        # ------------------------------------------------------------------
-        # Overall workspace background
-        root.configure(background="#f0f0f0")
-        # General widget colours
-        self.style.configure("TFrame", background="#f0f0f0")
-        self.style.configure("TLabel", background="#f0f0f0", foreground="black")
-        self.style.configure(
-            "TEntry", fieldbackground="#ffffff", background="#ffffff", foreground="black"
-        )
-        self.style.configure(
-            "TCombobox",
-            fieldbackground="#ffffff",
-            background="#ffffff",
-            foreground="black",
-        )
-        self.style.configure(
-            "TMenubutton", background="#e7edf5", foreground="black"
-        )
-        self.style.configure(
-            "TScrollbar",
-            background="#c0d4eb",
-            troughcolor="#e2e6eb",
-            bordercolor="#888888",
-            arrowcolor="#555555",
-            lightcolor="#eaf2fb",
-            darkcolor="#5a6d84",
-            borderwidth=2,
-            relief="raised",
-        )
-        # Apply the scrollbar styling to both orientations
-        for orient in ("Horizontal.TScrollbar", "Vertical.TScrollbar"):
-            self.style.configure(orient,
-                                background="#c0d4eb",
-                                troughcolor="#e2e6eb",
-                                bordercolor="#888888",
-                                arrowcolor="#555555",
-                                lightcolor="#eaf2fb",
-                                darkcolor="#5a6d84",
-                                borderwidth=2,
-                                relief="raised")
-        # Toolbox/LabelFrame titles
-        self.style.configure(
-            "Toolbox.TLabelframe",
-            background="#fef9e7",
-            bordercolor="#888888",
-            lightcolor="#fffef7",
-            darkcolor="#bfae6a",
-            borderwidth=1,
-            relief="raised",
-        )
-        self.style.configure(
-            "Toolbox.TLabelframe.Label",
-            background="#fef9e7",
-            foreground="black",
-            font=("Segoe UI", 10, "bold"),
-            padding=(4, 0, 0, 0),
-            anchor="w",
-        )
-        # Notebook (ribbon-like) title bars with beveled edges
-        self.style.configure(
-            "TNotebook",
-            background="#c0d4eb",
-            lightcolor="#eaf2fb",
-            darkcolor="#5a6d84",
-            borderwidth=2,
-            relief="raised",
-        )
-        self.style.configure(
-            "TNotebook.Tab",
-            background="#b5bdc9",
-            foreground="#555555",
-            borderwidth=1,
-            relief="raised",
-        )
-        self.style.map(
-            "TNotebook.Tab",
-            background=[("selected", "#4a6ea9"), ("!selected", "#b5bdc9")],
-            foreground=[("selected", "white"), ("!selected", "#555555")],
-        )
-        # Closable notebook shares the same appearance
-        self.style.configure(
-            "ClosableNotebook",
-            background="#c0d4eb",
-            lightcolor="#eaf2fb",
-            darkcolor="#5a6d84",
-            borderwidth=2,
-            relief="raised",
-        )
-        self.style.configure(
-            "ClosableNotebook.Tab",
-            background="#b5bdc9",
-            foreground="#555555",
-            borderwidth=1,
-            relief="raised",
-        )
-        self.style.map(
-            "ClosableNotebook.Tab",
-            background=[("selected", "#4a6ea9"), ("!selected", "#b5bdc9")],
-            foreground=[("selected", "white"), ("!selected", "#555555")],
-        )
-        # Mac-like capsule buttons
-        def _build_pill(top: str, bottom: str) -> tk.PhotoImage:
-            img = tk.PhotoImage(width=40, height=20)
-            # ``PhotoImage.put`` only accepts RGB colors. Some Tk builds
-            # (notably older Windows releases) mis-handle 8‑digit hex
-            # values used for transparency and instead raise a
-            # ``TclError``.  To keep the routine portable we fill the
-            # image with a solid RGB color and, where supported, mark that
-            # color as transparent.
-            img.put("#000000", to=(0, 0, 40, 20))
-            try:
-                img.transparency_set(0, 0, 0)
-            except Exception:
-                pass
-            radius = 10
-            t_r = int(top[1:3], 16)
-            t_g = int(top[3:5], 16)
-            t_b = int(top[5:7], 16)
-            b_r = int(bottom[1:3], 16)
-            b_g = int(bottom[3:5], 16)
-            b_b = int(bottom[5:7], 16)
-            for y in range(20):
-                ratio = y / 19
-                r = int(t_r * (1 - ratio) + b_r * ratio)
-                g = int(t_g * (1 - ratio) + b_g * ratio)
-                b = int(t_b * (1 - ratio) + b_b * ratio)
-                color = f"#{r:02x}{g:02x}{b:02x}"
-                for x in range(40):
-                    if x < radius:
-                        if (x - radius) ** 2 + (y - radius) ** 2 <= radius ** 2:
-                            img.put(color, (x, y))
-                    elif x >= 40 - radius:
-                        if (x - (40 - radius - 1)) ** 2 + (y - radius) ** 2 <= radius ** 2:
-                            img.put(color, (x, y))
-                    else:
-                        img.put(color, (x, y))
-            return img
-
-        self._btn_imgs = {
-            "normal": _build_pill("#fdfdfd", "#d2d2d2"),
-            "active": _build_pill("#eaeaea", "#c8c8c8"),
-            "pressed": _build_pill("#d0d0d0", "#a5a5a5"),
-        }
-        self.style.element_create(
-            "RoundedButton",
-            "image",
-            self._btn_imgs["normal"],
-            ("active", self._btn_imgs["active"]),
-            ("pressed", self._btn_imgs["pressed"]),
-            border=10,
-            sticky="nsew",
-        )
-        self.style.map(
-            "TButton",
-            relief=[("pressed", "sunken"), ("!pressed", "raised")],
-        )
-        # Navigation buttons used to scroll document tabs
+        self.style_app = StyleSubApp(root, self.style)
+        self.style_app.apply()
+        self._btn_imgs = self.style_app.btn_images
         self._init_nav_button_style()
-        # Increase notebook tab font/size so titles are fully visible
-        self.style.configure(
-            "TNotebook.Tab", font=("Arial", 10), padding=(10, 5), width=20
-        )
-        self.style.configure(
-            "ClosableNotebook.Tab", font=("Arial", 10), padding=(10, 5), width=20
-        )
+        self.tree_app = TreeSubApp()
+        self.fta_app = FTASubApp()
+        self.risk_app = RiskAssessmentSubApp()
+        self.reliability_app = ReliabilitySubApp()
+        self.helper = AutoML_Helper
         # style-aware icons used across tree views
         style_mgr = StyleManager.get_instance()
 
@@ -969,11 +762,17 @@ class AutoMLApp:
         self.user_manager = UserManager(self)
         self.project_manager = ProjectManager(self)
         self.cyber_manager = CyberSecurityManager(self)
-        self.diagram_controller = DiagramController(self)
+        self.diagram_export_app = DiagramExportSubApp(self)
+        self.use_case_diagram_app = UseCaseDiagramSubApp(self)
+        self.activity_diagram_app = ActivityDiagramSubApp(self)
+        self.block_diagram_app = BlockDiagramSubApp(self)
+        self.internal_block_diagram_app = InternalBlockDiagramSubApp(self)
+        self.control_flow_diagram_app = ControlFlowDiagramSubApp(self)
         self.sotif_manager = SOTIFManager(self)
         self.fmeda_manager = FMEDAManager(self)
         self.fmea_service = FMEAService(self)
         self.cta_manager = ControlTreeManager(self)
+        self.requirements_manager = RequirementsManagerSubApp(self)
 
         self.mechanism_libraries = []
         self.selected_mechanism_libraries = []
@@ -1094,11 +893,11 @@ class AutoMLApp:
         review_menu.add_command(label="Merge Review Comments", command=self.merge_review_comments)
         review_menu.add_command(label="Compare Versions", command=self.compare_versions)
         architecture_menu = tk.Menu(menubar, tearoff=0)
-        architecture_menu.add_command(label="Use Case Diagram", command=self.diagram_controller.open_use_case_diagram)
-        architecture_menu.add_command(label="Activity Diagram", command=self.diagram_controller.open_activity_diagram)
-        architecture_menu.add_command(label="Block Diagram", command=self.diagram_controller.open_block_diagram)
-        architecture_menu.add_command(label="Internal Block Diagram", command=self.diagram_controller.open_internal_block_diagram)
-        architecture_menu.add_command(label="Control Flow Diagram", command=self.diagram_controller.open_control_flow_diagram)
+        architecture_menu.add_command(label="Use Case Diagram", command=self.open_use_case_diagram)
+        architecture_menu.add_command(label="Activity Diagram", command=self.open_activity_diagram)
+        architecture_menu.add_command(label="Block Diagram", command=self.open_block_diagram)
+        architecture_menu.add_command(label="Internal Block Diagram", command=self.open_internal_block_diagram)
+        architecture_menu.add_command(label="Control Flow Diagram", command=self.open_control_flow_diagram)
         architecture_menu.add_separator()
         architecture_menu.add_command(
             label="AutoML Explorer",
@@ -1686,878 +1485,33 @@ class AutoMLApp:
         """Delegate to the FMEA service to display the FMEA manager."""
         self.fmea_service.show_fmea_list()
 
-    # --- Requirement Traceability Helpers used by reviews and matrix view ---
+        # --- Requirement Traceability Helpers used by reviews and matrix view ---
     def get_requirement_allocation_names(self, req_id):
-        """Return names of model elements linked to the requirement."""
-        names = []
-        repo = SysMLRepository.get_instance()
-        for diag_id, obj_id in repo.find_requirements(req_id):
-            diag = repo.diagrams.get(diag_id)
-            obj = next((o for o in getattr(diag, "objects", []) if o.get("obj_id") == obj_id), None)
-            dname = diag.name if diag else ""
-            oname = obj.get("properties", {}).get("name", "") if obj else ""
-            if dname and oname:
-                names.append(f"{dname}:{oname}")
-            elif dname or oname:
-                names.append(dname or oname)
-        for n in self.get_all_nodes(self.root_node):
-            reqs = getattr(n, "safety_requirements", [])
-            if any((r.get("id") if isinstance(r, dict) else getattr(r, "id", None)) == req_id for r in reqs):
-                names.append(n.user_name or f"Node {n.unique_id}")
-        for fmea in self.fmeas:
-            for e in fmea.get("entries", []):
-                reqs = e.get("safety_requirements", []) if isinstance(e, dict) else getattr(e, "safety_requirements", [])
-                if any((r.get("id") if isinstance(r, dict) else getattr(r, "id", None)) == req_id for r in reqs):
-                    if isinstance(e, dict):
-                        name = e.get("description") or e.get("user_name", f"BE {e.get('unique_id','')}")
-                    else:
-                        name = getattr(e, "description", "") or getattr(e, "user_name", f"BE {getattr(e, 'unique_id', '')}")
-                    names.append(f"{fmea['name']}:{name}")
-        repo = SysMLRepository.get_instance()
-        for diag in repo.diagrams.values():
-            for obj in getattr(diag, "objects", []):
-                reqs = obj.get("requirements", [])
-                if any(r.get("id") == req_id for r in reqs):
-                    name = obj.get("properties", {}).get("name") or obj.get("obj_type", "")
-                    names.append(name)
-        return names
-
-    def _collect_goal_names(self, node, acc):
-        if node.node_type.upper() == "TOP EVENT":
-            acc.add(node.safety_goal_description or (node.user_name or f"SG {node.unique_id}"))
-        for p in getattr(node, "parents", []):
-            self._collect_goal_names(p, acc)
+        return self.requirements_manager.get_requirement_allocation_names(req_id)
 
     def get_requirement_goal_names(self, req_id):
-        """Return a list of safety goal names linked to the requirement."""
-        goals = set()
-        for n in self.get_all_nodes(self.root_node):
-            reqs = getattr(n, "safety_requirements", [])
-            if any((r.get("id") if isinstance(r, dict) else getattr(r, "id", None)) == req_id for r in reqs):
-                self._collect_goal_names(n, goals)
-        for fmea in self.fmeas:
-            for e in fmea.get("entries", []):
-                reqs = e.get("safety_requirements", []) if isinstance(e, dict) else getattr(e, "safety_requirements", [])
-                if any((r.get("id") if isinstance(r, dict) else getattr(r, "id", None)) == req_id for r in reqs):
-                    parent_list = e.get("parents") if isinstance(e, dict) else getattr(e, "parents", None)
-                    parent = parent_list[0] if parent_list else None
-                    if isinstance(parent, dict) and "unique_id" in parent:
-                        node = self.find_node_by_id_all(parent["unique_id"])
-                    else:
-                        node = parent if hasattr(parent, "unique_id") else None
-                    if node:
-                        self._collect_goal_names(node, goals)
-        return sorted(goals)
+        return self.requirements_manager.get_requirement_goal_names(req_id)
 
     def format_requirement_with_trace(self, req):
-        """Return requirement text including allocation and safety goal lists."""
-        rid = req.get("id", "")
-        alloc = ", ".join(self.get_requirement_allocation_names(rid))
-        goals = ", ".join(self.get_requirement_goal_names(rid))
-        base = format_requirement(req)
-        return f"{base} (Alloc: {alloc}; SGs: {goals})"
+        return self.requirements_manager.format_requirement_with_trace(req)
 
     def build_requirement_diff_html(self, review):
-        """Return HTML highlighting requirement differences for the review."""
-        if not self.versions:
-            return ""
-        base_data = self.versions[-1]["data"]
-        current = self.export_model_data(include_versions=False)
-
-        def filter_data(data):
-            return {
-                "top_events": [t for t in data.get("top_events", []) if t["unique_id"] in review.fta_ids],
-                "fmeas": [f for f in data.get("fmeas", []) if f["name"] in review.fmea_names],
-                "fmedas": [d for d in data.get("fmedas", []) if d.get("name") in getattr(review, "fmeda_names", [])],
-            }
-
-        data1 = filter_data(base_data)
-        data2 = filter_data(current)
-        map1 = self.node_map_from_data(data1["top_events"])
-        map2 = self.node_map_from_data(data2["top_events"])
-
-        def collect_reqs(node_dict, target):
-            for r in node_dict.get("safety_requirements", []):
-                rid = r.get("id")
-                if rid and rid not in target:
-                    target[rid] = r
-            for ch in node_dict.get("children", []):
-                collect_reqs(ch, target)
-
-        reqs1, reqs2 = {}, {}
-        for nid in review.fta_ids:
-            if nid in map1:
-                collect_reqs(map1[nid], reqs1)
-            if nid in map2:
-                collect_reqs(map2[nid], reqs2)
-
-        fmea1 = {f["name"]: f for f in data1.get("fmeas", [])}
-        fmea2 = {f["name"]: f for f in data2.get("fmeas", [])}
-        for name in review.fmea_names:
-            for e in fmea1.get(name, {}).get("entries", []):
-                for r in e.get("safety_requirements", []):
-                    rid = r.get("id")
-                    if rid and rid not in reqs1:
-                        reqs1[rid] = r
-            for e in fmea2.get(name, {}).get("entries", []):
-                for r in e.get("safety_requirements", []):
-                    rid = r.get("id")
-                    if rid and rid not in reqs2:
-                        reqs2[rid] = r
-        for f in data1.get("fmedas", []):
-            for e in f.get("entries", []):
-                for r in e.get("safety_requirements", []):
-                    rid = r.get("id")
-                    if rid and rid not in reqs1:
-                        reqs1[rid] = r
-        for f in data2.get("fmedas", []):
-            for e in f.get("entries", []):
-                for r in e.get("safety_requirements", []):
-                    rid = r.get("id")
-                    if rid and rid not in reqs2:
-                        reqs2[rid] = r
-        for f in data1.get("fmedas", []):
-            for e in f.get("entries", []):
-                for r in e.get("safety_requirements", []):
-                    rid = r.get("id")
-                    if rid and rid not in reqs1:
-                        reqs1[rid] = r
-        for f in data2.get("fmedas", []):
-            for e in f.get("entries", []):
-                for r in e.get("safety_requirements", []):
-                    rid = r.get("id")
-                    if rid and rid not in reqs2:
-                        reqs2[rid] = r
-        for f in data1.get("fmedas", []):
-            for e in f.get("entries", []):
-                for r in e.get("safety_requirements", []):
-                    rid = r.get("id")
-                    if rid and rid not in reqs1:
-                        reqs1[rid] = r
-        for f in data2.get("fmedas", []):
-            for e in f.get("entries", []):
-                for r in e.get("safety_requirements", []):
-                    rid = r.get("id")
-                    if rid and rid not in reqs2:
-                        reqs2[rid] = r
-        for f in data1.get("fmedas", []):
-            for e in f.get("entries", []):
-                for r in e.get("safety_requirements", []):
-                    rid = r.get("id")
-                    if rid and rid not in reqs1:
-                        reqs1[rid] = r
-        for f in data2.get("fmedas", []):
-            for e in f.get("entries", []):
-                for r in e.get("safety_requirements", []):
-                    rid = r.get("id")
-                    if rid and rid not in reqs2:
-                        reqs2[rid] = r
-        for f in data1.get("fmedas", []):
-            for e in f.get("entries", []):
-                for r in e.get("safety_requirements", []):
-                    rid = r.get("id")
-                    if rid and rid not in reqs1:
-                        reqs1[rid] = r
-        for f in data2.get("fmedas", []):
-            for e in f.get("entries", []):
-                for r in e.get("safety_requirements", []):
-                    rid = r.get("id")
-                    if rid and rid not in reqs2:
-                        reqs2[rid] = r
-        for f in data1.get("fmedas", []):
-            for e in f.get("entries", []):
-                for r in e.get("safety_requirements", []):
-                    rid = r.get("id")
-                    if rid and rid not in reqs1:
-                        reqs1[rid] = r
-        for f in data2.get("fmedas", []):
-            for e in f.get("entries", []):
-                for r in e.get("safety_requirements", []):
-                    rid = r.get("id")
-                    if rid and rid not in reqs2:
-                        reqs2[rid] = r
-        for f in data1.get("fmedas", []):
-            for e in f.get("entries", []):
-                for r in e.get("safety_requirements", []):
-                    rid = r.get("id")
-                    if rid and rid not in reqs1:
-                        reqs1[rid] = r
-        for f in data2.get("fmedas", []):
-            for e in f.get("entries", []):
-                for r in e.get("safety_requirements", []):
-                    rid = r.get("id")
-                    if rid and rid not in reqs2:
-                        reqs2[rid] = r
-
-        import difflib, html
-
-        def html_diff(a, b):
-            matcher = difflib.SequenceMatcher(None, a, b)
-            parts = []
-            for tag, i1, i2, j1, j2 in matcher.get_opcodes():
-                if tag == "equal":
-                    parts.append(html.escape(a[i1:i2]))
-                elif tag == "delete":
-                    parts.append(f"<span style='color:red'>{html.escape(a[i1:i2])}</span>")
-                elif tag == "insert":
-                    parts.append(f"<span style='color:blue'>{html.escape(b[j1:j2])}</span>")
-                elif tag == "replace":
-                    parts.append(f"<span style='color:red'>{html.escape(a[i1:i2])}</span>")
-                    parts.append(f"<span style='color:blue'>{html.escape(b[j1:j2])}</span>")
-            return "".join(parts)
-
-        lines = []
-        all_ids = sorted(set(reqs1) | set(reqs2))
-        for rid in all_ids:
-            r1 = reqs1.get(rid)
-            r2 = reqs2.get(rid)
-            if r1 and not r2:
-                lines.append(f"Removed: {html.escape(self.format_requirement_with_trace(r1))}")
-            elif r2 and not r1:
-                lines.append(f"Added: {html.escape(self.format_requirement_with_trace(r2))}")
-            else:
-                if json.dumps(r1, sort_keys=True) != json.dumps(r2, sort_keys=True):
-                    lines.append("Updated: " + html_diff(self.format_requirement_with_trace(r1), self.format_requirement_with_trace(r2)))
-
-        for nid in review.fta_ids:
-            n1 = map1.get(nid, {})
-            n2 = map2.get(nid, {})
-            sg_old = f"{n1.get('safety_goal_description','')} [{n1.get('safety_goal_asil','')}]"
-            sg_new = f"{n2.get('safety_goal_description','')} [{n2.get('safety_goal_asil','')}]"
-            label = n2.get('user_name') or n1.get('user_name') or f"Node {nid}"
-            if sg_old != sg_new:
-                lines.append(
-                    f"Safety Goal for {html.escape(label)}: " + html_diff(sg_old, sg_new)
-                )
-            if n1.get('safe_state','') != n2.get('safe_state',''):
-                lines.append(
-                    f"Safe State for {html.escape(label)}: " + html_diff(n1.get('safe_state',''), n2.get('safe_state',''))
-                )
-
-        return "<br>".join(lines)
-
-    # --- Requirement Traceability Helpers used by reviews and matrix view ---
-    def get_requirement_allocation_names(self, req_id):
-        """Return a list of node or FMEA entry names where the requirement appears."""
-        names = []
-        repo = SysMLRepository.get_instance()
-        for diag_id, obj_id in repo.find_requirements(req_id):
-            diag = repo.diagrams.get(diag_id)
-            obj = next((o for o in getattr(diag, "objects", []) if o.get("obj_id") == obj_id), None)
-            dname = diag.name if diag else ""
-            oname = obj.get("properties", {}).get("name", "") if obj else ""
-            if dname and oname:
-                names.append(f"{dname}:{oname}")
-            elif dname or oname:
-                names.append(dname or oname)
-        for n in self.get_all_nodes(self.root_node):
-            reqs = getattr(n, "safety_requirements", [])
-            if any((r.get("id") if isinstance(r, dict) else getattr(r, "id", None)) == req_id for r in reqs):
-                names.append(n.user_name or f"Node {n.unique_id}")
-        for fmea in self.fmeas:
-            for e in fmea.get("entries", []):
-                reqs = e.get("safety_requirements", []) if isinstance(e, dict) else getattr(e, "safety_requirements", [])
-                if any((r.get("id") if isinstance(r, dict) else getattr(r, "id", None)) == req_id for r in reqs):
-                    if isinstance(e, dict):
-                        name = e.get("description") or e.get("user_name", f"BE {e.get('unique_id','')}")
-                    else:
-                        name = getattr(e, "description", "") or getattr(e, "user_name", f"BE {getattr(e, 'unique_id', '')}")
-                    names.append(f"{fmea['name']}:{name}")
-        return names
-
-    def _collect_goal_names(self, node, acc):
-        if node.node_type.upper() == "TOP EVENT":
-            acc.add(node.safety_goal_description or (node.user_name or f"SG {node.unique_id}"))
-        for p in getattr(node, "parents", []):
-            self._collect_goal_names(p, acc)
-
-    def get_requirement_goal_names(self, req_id):
-        """Return a list of safety goal names linked to the requirement."""
-        goals = set()
-        for n in self.get_all_nodes(self.root_node):
-            reqs = getattr(n, "safety_requirements", [])
-            if any((r.get("id") if isinstance(r, dict) else getattr(r, "id", None)) == req_id for r in reqs):
-                self._collect_goal_names(n, goals)
-        for fmea in self.fmeas:
-            for e in fmea.get("entries", []):
-                reqs = e.get("safety_requirements", []) if isinstance(e, dict) else getattr(e, "safety_requirements", [])
-                if any((r.get("id") if isinstance(r, dict) else getattr(r, "id", None)) == req_id for r in reqs):
-                    parent_list = e.get("parents", []) if isinstance(e, dict) else getattr(e, "parents", [])
-                    parent = parent_list[0] if parent_list else None
-                    if isinstance(parent, dict) and "unique_id" in parent:
-                        node = self.find_node_by_id_all(parent["unique_id"])
-                    else:
-                        node = parent if hasattr(parent, "unique_id") else None
-                    if node:
-                        self._collect_goal_names(node, goals)
-        return sorted(goals)
-
-    def format_requirement_with_trace(self, req):
-        """Return requirement text including allocation and safety goal lists."""
-        rid = req.get("id", "")
-        alloc = ", ".join(self.get_requirement_allocation_names(rid))
-        goals = ", ".join(self.get_requirement_goal_names(rid))
-        base = format_requirement(req)
-        return f"{base} (Alloc: {alloc}; SGs: {goals})"
-
-    def build_requirement_diff_html(self, review):
-        """Return HTML highlighting requirement differences for the review."""
-        if not self.versions:
-            return ""
-        base_data = self.versions[-1]["data"]
-        current = self.export_model_data(include_versions=False)
-
-        def filter_data(data):
-            return {
-                "top_events": [t for t in data.get("top_events", []) if t["unique_id"] in review.fta_ids],
-                "fmeas": [f for f in data.get("fmeas", []) if f["name"] in review.fmea_names],
-                "fmedas": [d for d in data.get("fmedas", []) if d.get("name") in getattr(review, "fmeda_names", [])],
-            }
-
-        data1 = filter_data(base_data)
-        data2 = filter_data(current)
-        map1 = self.node_map_from_data(data1["top_events"])
-        map2 = self.node_map_from_data(data2["top_events"])
-
-        def collect_reqs(node_dict, target):
-            for r in node_dict.get("safety_requirements", []):
-                rid = r.get("id")
-                if rid and rid not in target:
-                    target[rid] = r
-            for ch in node_dict.get("children", []):
-                collect_reqs(ch, target)
-
-        reqs1, reqs2 = {}, {}
-        for nid in review.fta_ids:
-            if nid in map1:
-                collect_reqs(map1[nid], reqs1)
-            if nid in map2:
-                collect_reqs(map2[nid], reqs2)
-
-        fmea1 = {f["name"]: f for f in data1.get("fmeas", [])}
-        fmea2 = {f["name"]: f for f in data2.get("fmeas", [])}
-        for name in review.fmea_names:
-            for e in fmea1.get(name, {}).get("entries", []):
-                for r in e.get("safety_requirements", []):
-                    rid = r.get("id")
-                    if rid and rid not in reqs1:
-                        reqs1[rid] = r
-            for e in fmea2.get(name, {}).get("entries", []):
-                for r in e.get("safety_requirements", []):
-                    rid = r.get("id")
-                    if rid and rid not in reqs2:
-                        reqs2[rid] = r
-
-        import difflib, html
-
-        def html_diff(a, b):
-            matcher = difflib.SequenceMatcher(None, a, b)
-            parts = []
-            for tag, i1, i2, j1, j2 in matcher.get_opcodes():
-                if tag == "equal":
-                    parts.append(html.escape(a[i1:i2]))
-                elif tag == "delete":
-                    parts.append(f"<span style='color:red'>{html.escape(a[i1:i2])}</span>")
-                elif tag == "insert":
-                    parts.append(f"<span style='color:blue'>{html.escape(b[j1:j2])}</span>")
-                elif tag == "replace":
-                    parts.append(f"<span style='color:red'>{html.escape(a[i1:i2])}</span>")
-                    parts.append(f"<span style='color:blue'>{html.escape(b[j1:j2])}</span>")
-            return "".join(parts)
-
-        lines = []
-        all_ids = sorted(set(reqs1) | set(reqs2))
-        for rid in all_ids:
-            r1 = reqs1.get(rid)
-            r2 = reqs2.get(rid)
-            if r1 and not r2:
-                lines.append(f"Removed: {html.escape(self.format_requirement_with_trace(r1))}")
-            elif r2 and not r1:
-                lines.append(f"Added: {html.escape(self.format_requirement_with_trace(r2))}")
-            else:
-                if json.dumps(r1, sort_keys=True) != json.dumps(r2, sort_keys=True):
-                    lines.append("Updated: " + html_diff(self.format_requirement_with_trace(r1), self.format_requirement_with_trace(r2)))
-
-        for nid in review.fta_ids:
-            n1 = map1.get(nid, {})
-            n2 = map2.get(nid, {})
-            sg_old = f"{n1.get('safety_goal_description','')} [{n1.get('safety_goal_asil','')}]"
-            sg_new = f"{n2.get('safety_goal_description','')} [{n2.get('safety_goal_asil','')}]"
-            label = n2.get('user_name') or n1.get('user_name') or f"Node {nid}"
-            if sg_old != sg_new:
-                lines.append(
-                    f"Safety Goal for {html.escape(label)}: " + html_diff(sg_old, sg_new)
-                )
-            if n1.get('safe_state','') != n2.get('safe_state',''):
-                lines.append(
-                    f"Safe State for {html.escape(label)}: " + html_diff(n1.get('safe_state',''), n2.get('safe_state',''))
-                )
-
-        return "<br>".join(lines)
-
-    # --- Requirement Traceability Helpers used by reviews and matrix view ---
-    def get_requirement_allocation_names(self, req_id):
-        """Return a list of node or FMEA entry names where the requirement appears."""
-        names = []
-        repo = SysMLRepository.get_instance()
-        for diag_id, obj_id in repo.find_requirements(req_id):
-            diag = repo.diagrams.get(diag_id)
-            obj = next((o for o in getattr(diag, "objects", []) if o.get("obj_id") == obj_id), None)
-            dname = diag.name if diag else ""
-            oname = obj.get("properties", {}).get("name", "") if obj else ""
-            if dname and oname:
-                names.append(f"{dname}:{oname}")
-            elif dname or oname:
-                names.append(dname or oname)
-        for n in self.get_all_nodes(self.root_node):
-            reqs = getattr(n, "safety_requirements", [])
-            if any((r.get("id") if isinstance(r, dict) else getattr(r, "id", None)) == req_id for r in reqs):
-                names.append(n.user_name or f"Node {n.unique_id}")
-        for fmea in self.fmeas:
-            for e in fmea.get("entries", []):
-                reqs = e.get("safety_requirements", []) if isinstance(e, dict) else getattr(e, "safety_requirements", [])
-                if any((r.get("id") if isinstance(r, dict) else getattr(r, "id", None)) == req_id for r in reqs):
-                    if isinstance(e, dict):
-                        name = e.get("description") or e.get("user_name", f"BE {e.get('unique_id','')}")
-                    else:
-                        name = getattr(e, "description", "") or getattr(e, "user_name", f"BE {getattr(e, 'unique_id', '')}")
-                    names.append(f"{fmea['name']}:{name}")
-        return names
-
-    def _collect_goal_names(self, node, acc):
-        if node.node_type.upper() == "TOP EVENT":
-            acc.add(node.safety_goal_description or (node.user_name or f"SG {node.unique_id}"))
-        for p in getattr(node, "parents", []):
-            self._collect_goal_names(p, acc)
-
-    def get_requirement_goal_names(self, req_id):
-        """Return a list of safety goal names linked to the requirement."""
-        goals = set()
-        for n in self.get_all_nodes(self.root_node):
-            reqs = getattr(n, "safety_requirements", [])
-            if any((r.get("id") if isinstance(r, dict) else getattr(r, "id", None)) == req_id for r in reqs):
-                self._collect_goal_names(n, goals)
-        for fmea in self.fmeas:
-            for e in fmea.get("entries", []):
-                reqs = e.get("safety_requirements", []) if isinstance(e, dict) else getattr(e, "safety_requirements", [])
-                if any((r.get("id") if isinstance(r, dict) else getattr(r, "id", None)) == req_id for r in reqs):
-                    parent_list = e.get("parents", []) if isinstance(e, dict) else getattr(e, "parents", [])
-                    parent = parent_list[0] if parent_list else None
-                    if isinstance(parent, dict) and "unique_id" in parent:
-                        node = self.find_node_by_id_all(parent["unique_id"])
-                    else:
-                        node = parent if hasattr(parent, "unique_id") else None
-                    if node:
-                        self._collect_goal_names(node, goals)
-        return sorted(goals)
-
-    def format_requirement_with_trace(self, req):
-        """Return requirement text including allocation and safety goal lists."""
-        if isinstance(req, dict):
-            data = req
-        else:
-            data = {
-                "id": getattr(req, "id", ""),
-                "req_type": getattr(req, "req_type", ""),
-                "asil": getattr(req, "asil", ""),
-                "cal": getattr(req, "cal", ""),
-                "text": getattr(req, "text", ""),
-            }
-        rid = data.get("id", "")
-        alloc = ", ".join(self.get_requirement_allocation_names(rid))
-        goals = ", ".join(self.get_requirement_goal_names(rid))
-        base = format_requirement(data)
-        return f"{base} (Alloc: {alloc}; SGs: {goals})"
-
-    def build_requirement_diff_html(self, review):
-        """Return HTML highlighting requirement differences for the review."""
-        if not self.versions:
-            return ""
-        base_data = self.versions[-1]["data"]
-        current = self.export_model_data(include_versions=False)
-
-        def filter_data(data):
-            return {
-                "top_events": [
-                    t for t in data.get("top_events", []) if t["unique_id"] in review.fta_ids
-                ],
-                "fmeas": [
-                    f for f in data.get("fmeas", []) if f["name"] in review.fmea_names
-                ],
-                "fmedas": [
-                    d
-                    for d in data.get("fmedas", [])
-                    if d.get("name") in getattr(review, "fmeda_names", [])
-                ],
-                "hazops": [
-                    d
-                    for d in data.get("hazops", [])
-                    if d.get("name") in getattr(review, "hazop_names", [])
-                ],
-                "haras": [
-                    d
-                    for d in data.get("haras", [])
-                    if d.get("name") in getattr(review, "hara_names", [])
-                ],
-                "stpas": [
-                    d
-                    for d in data.get("stpas", [])
-                    if d.get("name") in getattr(review, "stpa_names", [])
-                ],
-            }
-
-        data1 = filter_data(base_data)
-        data2 = filter_data(current)
-        map1 = self.node_map_from_data(data1["top_events"])
-        map2 = self.node_map_from_data(data2["top_events"])
-
-        def collect_reqs(node_dict, target):
-            for r in node_dict.get("safety_requirements", []):
-                rid = r.get("id")
-                if rid and rid not in target:
-                    target[rid] = r
-            for ch in node_dict.get("children", []):
-                collect_reqs(ch, target)
-
-        reqs1, reqs2 = {}, {}
-        for nid in review.fta_ids:
-            if nid in map1:
-                collect_reqs(map1[nid], reqs1)
-            if nid in map2:
-                collect_reqs(map2[nid], reqs2)
-
-        fmea1 = {f["name"]: f for f in data1.get("fmeas", [])}
-        fmea2 = {f["name"]: f for f in data2.get("fmeas", [])}
-        for name in review.fmea_names:
-            for e in fmea1.get(name, {}).get("entries", []):
-                for r in e.get("safety_requirements", []):
-                    rid = r.get("id")
-                    if rid and rid not in reqs1:
-                        reqs1[rid] = r
-            for e in fmea2.get(name, {}).get("entries", []):
-                for r in e.get("safety_requirements", []):
-                    rid = r.get("id")
-                    if rid and rid not in reqs2:
-                        reqs2[rid] = r
-
-        import difflib, html
-
-        def html_diff(a, b):
-            matcher = difflib.SequenceMatcher(None, a, b)
-            parts = []
-            for tag, i1, i2, j1, j2 in matcher.get_opcodes():
-                if tag == "equal":
-                    parts.append(html.escape(a[i1:i2]))
-                elif tag == "delete":
-                    parts.append(f"<span style='color:red'>{html.escape(a[i1:i2])}</span>")
-                elif tag == "insert":
-                    parts.append(f"<span style='color:blue'>{html.escape(b[j1:j2])}</span>")
-                elif tag == "replace":
-                    parts.append(f"<span style='color:red'>{html.escape(a[i1:i2])}</span>")
-                    parts.append(f"<span style='color:blue'>{html.escape(b[j1:j2])}</span>")
-            return "".join(parts)
-
-        lines = []
-        all_ids = sorted(set(reqs1) | set(reqs2))
-        for rid in all_ids:
-            r1 = reqs1.get(rid)
-            r2 = reqs2.get(rid)
-            if r1 and not r2:
-                lines.append(f"Removed: {html.escape(self.format_requirement_with_trace(r1))}")
-            elif r2 and not r1:
-                lines.append(f"Added: {html.escape(self.format_requirement_with_trace(r2))}")
-            else:
-                if json.dumps(r1, sort_keys=True) != json.dumps(r2, sort_keys=True):
-                    lines.append("Updated: " + html_diff(self.format_requirement_with_trace(r1), self.format_requirement_with_trace(r2)))
-
-        for nid in review.fta_ids:
-            n1 = map1.get(nid, {})
-            n2 = map2.get(nid, {})
-            sg_old = f"{n1.get('safety_goal_description','')} [{n1.get('safety_goal_asil','')}]"
-            sg_new = f"{n2.get('safety_goal_description','')} [{n2.get('safety_goal_asil','')}]"
-            label = n2.get('user_name') or n1.get('user_name') or f"Node {nid}"
-            if sg_old != sg_new:
-                lines.append(
-                    f"Safety Goal for {html.escape(label)}: " + html_diff(sg_old, sg_new)
-                )
-            if n1.get('safe_state','') != n2.get('safe_state',''):
-                lines.append(
-                    f"Safe State for {html.escape(label)}: " + html_diff(n1.get('safe_state',''), n2.get('safe_state',''))
-                )
-
-        return "<br>".join(lines)
-
-    # --- Requirement Traceability Helpers used by reviews and matrix view ---
-    def get_requirement_allocation_names(self, req_id):
-        """Return a list of node or FMEA entry names where the requirement appears."""
-        names = []
-        repo = SysMLRepository.get_instance()
-        for diag_id, obj_id in repo.find_requirements(req_id):
-            diag = repo.diagrams.get(diag_id)
-            obj = next((o for o in getattr(diag, "objects", []) if o.get("obj_id") == obj_id), None)
-            dname = diag.name if diag else ""
-            oname = obj.get("properties", {}).get("name", "") if obj else ""
-            if dname and oname:
-                names.append(f"{dname}:{oname}")
-            elif dname or oname:
-                names.append(dname or oname)
-        for n in self.get_all_nodes(self.root_node):
-            reqs = getattr(n, "safety_requirements", [])
-            if any((r.get("id") if isinstance(r, dict) else getattr(r, "id", None)) == req_id for r in reqs):
-                names.append(n.user_name or f"Node {n.unique_id}")
-        for fmea in self.fmeas:
-            for e in fmea.get("entries", []):
-                reqs = e.get("safety_requirements", []) if isinstance(e, dict) else getattr(e, "safety_requirements", [])
-                if any((r.get("id") if isinstance(r, dict) else getattr(r, "id", None)) == req_id for r in reqs):
-                    if isinstance(e, dict):
-                        name = e.get("description") or e.get("user_name", f"BE {e.get('unique_id','')}")
-                    else:
-                        name = getattr(e, "description", "") or getattr(e, "user_name", f"BE {getattr(e, 'unique_id', '')}")
-                    names.append(f"{fmea['name']}:{name}")
-        return names
-
-    def _collect_goal_names(self, node, acc):
-        if node.node_type.upper() == "TOP EVENT":
-            acc.add(node.safety_goal_description or (node.user_name or f"SG {node.unique_id}"))
-        for p in getattr(node, "parents", []):
-            self._collect_goal_names(p, acc)
-
-    def get_requirement_goal_names(self, req_id):
-        """Return a list of safety goal names linked to the requirement."""
-        goals = set()
-        for n in self.get_all_nodes(self.root_node):
-            reqs = getattr(n, "safety_requirements", [])
-            if any((r.get("id") if isinstance(r, dict) else getattr(r, "id", None)) == req_id for r in reqs):
-                self._collect_goal_names(n, goals)
-        for fmea in self.fmeas:
-            for e in fmea.get("entries", []):
-                reqs = e.get("safety_requirements", []) if isinstance(e, dict) else getattr(e, "safety_requirements", [])
-                if any((r.get("id") if isinstance(r, dict) else getattr(r, "id", None)) == req_id for r in reqs):
-                    parent_list = e.get("parents", []) if isinstance(e, dict) else getattr(e, "parents", [])
-                    parent = parent_list[0] if parent_list else None
-                    if isinstance(parent, dict) and "unique_id" in parent:
-                        node = self.find_node_by_id_all(parent["unique_id"])
-                    else:
-                        node = parent if hasattr(parent, "unique_id") else None
-                    if node:
-                        self._collect_goal_names(node, goals)
-        return sorted(goals)
-
-    def format_requirement_with_trace(self, req):
-        """Return requirement text including allocation and safety goal lists."""
-        if isinstance(req, dict):
-            data = req
-        else:
-            data = {
-                "id": getattr(req, "id", ""),
-                "req_type": getattr(req, "req_type", ""),
-                "asil": getattr(req, "asil", ""),
-                "cal": getattr(req, "cal", ""),
-                "text": getattr(req, "text", ""),
-            }
-        rid = data.get("id", "")
-        alloc = ", ".join(self.get_requirement_allocation_names(rid))
-        goals = ", ".join(self.get_requirement_goal_names(rid))
-        base = format_requirement(data)
-        return f"{base} (Alloc: {alloc}; SGs: {goals})"
-
-    def build_requirement_diff_html(self, review):
-        """Return HTML highlighting requirement differences for the review."""
-        if not self.versions:
-            return ""
-        base_data = self.versions[-1]["data"]
-        current = self.export_model_data(include_versions=False)
-
-        def filter_data(data):
-            return {
-                "top_events": [
-                    t for t in data.get("top_events", []) if t["unique_id"] in review.fta_ids
-                ],
-                "fmeas": [f for f in data.get("fmeas", []) if f["name"] in review.fmea_names],
-                "fmedas": [
-                    d
-                    for d in data.get("fmedas", [])
-                    if d.get("name") in getattr(review, "fmeda_names", [])
-                ],
-                "hazops": [
-                    d
-                    for d in data.get("hazops", [])
-                    if d.get("name") in getattr(review, "hazop_names", [])
-                ],
-                "haras": [
-                    d
-                    for d in data.get("haras", [])
-                    if d.get("name") in getattr(review, "hara_names", [])
-                ],
-                "stpas": [
-                    d
-                    for d in data.get("stpas", [])
-                    if d.get("name") in getattr(review, "stpa_names", [])
-                ],
-            }
-
-        data1 = filter_data(base_data)
-        data2 = filter_data(current)
-        map1 = self.node_map_from_data(data1["top_events"])
-        map2 = self.node_map_from_data(data2["top_events"])
-
-        def collect_reqs(node_dict, target):
-            for r in node_dict.get("safety_requirements", []):
-                rid = r.get("id")
-                if rid and rid not in target:
-                    target[rid] = r
-            for ch in node_dict.get("children", []):
-                collect_reqs(ch, target)
-
-        reqs1, reqs2 = {}, {}
-        for nid in review.fta_ids:
-            if nid in map1:
-                collect_reqs(map1[nid], reqs1)
-            if nid in map2:
-                collect_reqs(map2[nid], reqs2)
-
-        fmea1 = {f["name"]: f for f in data1.get("fmeas", [])}
-        fmea2 = {f["name"]: f for f in data2.get("fmeas", [])}
-        for name in review.fmea_names:
-            for e in fmea1.get(name, {}).get("entries", []):
-                for r in e.get("safety_requirements", []):
-                    rid = r.get("id")
-                    if rid and rid not in reqs1:
-                        reqs1[rid] = r
-            for e in fmea2.get(name, {}).get("entries", []):
-                for r in e.get("safety_requirements", []):
-                    rid = r.get("id")
-                    if rid and rid not in reqs2:
-                        reqs2[rid] = r
-        for f in data1.get("fmedas", []):
-            for e in f.get("entries", []):
-                for r in e.get("safety_requirements", []):
-                    rid = r.get("id")
-                    if rid and rid not in reqs1:
-                        reqs1[rid] = r
-        for f in data2.get("fmedas", []):
-            for e in f.get("entries", []):
-                for r in e.get("safety_requirements", []):
-                    rid = r.get("id")
-                    if rid and rid not in reqs2:
-                        reqs2[rid] = r
-
-        import difflib, html
-
-        def html_diff(a, b):
-            matcher = difflib.SequenceMatcher(None, a, b)
-            parts = []
-            for tag, i1, i2, j1, j2 in matcher.get_opcodes():
-                if tag == "equal":
-                    parts.append(html.escape(a[i1:i2]))
-                elif tag == "delete":
-                    parts.append(f"<span style='color:red'>{html.escape(a[i1:i2])}</span>")
-                elif tag == "insert":
-                    parts.append(f"<span style='color:blue'>{html.escape(b[j1:j2])}</span>")
-                elif tag == "replace":
-                    parts.append(f"<span style='color:red'>{html.escape(a[i1:i2])}</span>")
-                    parts.append(f"<span style='color:blue'>{html.escape(b[j1:j2])}</span>")
-            return "".join(parts)
-
-        lines = []
-        all_ids = sorted(set(reqs1) | set(reqs2))
-        for rid in all_ids:
-            r1 = reqs1.get(rid)
-            r2 = reqs2.get(rid)
-            if r1 and not r2:
-                lines.append(f"Removed: {html.escape(self.format_requirement_with_trace(r1))}")
-            elif r2 and not r1:
-                lines.append(f"Added: {html.escape(self.format_requirement_with_trace(r2))}")
-            else:
-                if json.dumps(r1, sort_keys=True) != json.dumps(r2, sort_keys=True):
-                    lines.append("Updated: " + html_diff(self.format_requirement_with_trace(r1), self.format_requirement_with_trace(r2)))
-
-        for nid in review.fta_ids:
-            n1 = map1.get(nid, {})
-            n2 = map2.get(nid, {})
-            sg_old = f"{n1.get('safety_goal_description','')} [{n1.get('safety_goal_asil','')}]"
-            sg_new = f"{n2.get('safety_goal_description','')} [{n2.get('safety_goal_asil','')}]"
-            label = n2.get('user_name') or n1.get('user_name') or f"Node {nid}"
-            if sg_old != sg_new:
-                lines.append(
-                    f"Safety Goal for {html.escape(label)}: " + html_diff(sg_old, sg_new)
-                )
-            if n1.get('safe_state','') != n2.get('safe_state',''):
-                lines.append(
-                    f"Safe State for {html.escape(label)}: " + html_diff(n1.get('safe_state',''), n2.get('safe_state',''))
-                )
-
-        return "<br>".join(lines)
+        return self.requirements_manager.build_requirement_diff_html(review)
 
     def generate_recommendations_for_top_event(self, node):
-        # Determine the Prototype Assurance Level (PAL) based on the node’s quantitative score.
-        level = AutoML_Helper.discretize_level(node.quant_value) if node.quant_value is not None else 1
-        rec = dynamic_recommendations.get(level, {})
-        rec_text = f"<b>Recommendations for Prototype Assurance Level (PAL) {level}:</b><br/>"
-        for category in ["Testing Requirements", "IFTD Responsibilities", "Preventive Maintenance Actions", "Relevant AVSC Guidelines"]:
-            if category in rec:
-                rec_text += f"<b>{category}:</b><br/><ul><li>{rec[category]}</li></ul><br/>"
-        return rec_text
+        return self.fta_app.generate_recommendations_for_top_event(self, node)
 
     def back_all_pages(self):
-        if self.page_history:
-            # Jump to the very first page saved in history:
-            first_page = self.page_history[0]
-            # Clear the history so that subsequent back presses do not try to go further.
-            self.page_history = []
-            for widget in self.canvas_frame.winfo_children():
-                widget.destroy()
-            self.open_page_diagram(first_page)
-        else:
-            # No history: you could simply reinitialize the main diagram
-            self.close_page_diagram()
+        return self.fta_app.back_all_pages(self)
 
     def move_top_event_up(self):
-        sel = self.analysis_tree.selection()
-        if not sel:
-            messagebox.showwarning("Move Up", "Select a top-level event to move.")
-            return
-        try:
-            node_id = int(self.analysis_tree.item(sel[0], "tags")[0])
-        except Exception:
-            return
-        # Find the index in the top_events list.
-        index = next((i for i, event in enumerate(self.top_events) if event.unique_id == node_id), None)
-        if index is None:
-            messagebox.showwarning("Move Up", "The selected node is not a top-level event.")
-            return
-        if index == 0:
-            messagebox.showinfo("Move Up", "This event is already at the top.")
-            return
-        # Swap with the one above it.
-        self.top_events[index], self.top_events[index - 1] = self.top_events[index - 1], self.top_events[index]
-        self.update_views()
+        return self.fta_app.move_top_event_up(self)
 
     def move_top_event_down(self):
-        sel = self.analysis_tree.selection()
-        if not sel:
-            messagebox.showwarning("Move Down", "Select a top-level event to move.")
-            return
-        try:
-            node_id = int(self.analysis_tree.item(sel[0], "tags")[0])
-        except Exception:
-            return
-        index = next((i for i, event in enumerate(self.top_events) if event.unique_id == node_id), None)
-        if index is None:
-            messagebox.showwarning("Move Down", "The selected node is not a top-level event.")
-            return
-        if index == len(self.top_events) - 1:
-            messagebox.showinfo("Move Down", "This event is already at the bottom.")
-            return
-        # Swap with the one below it.
-        self.top_events[index], self.top_events[index + 1] = self.top_events[index + 1], self.top_events[index]
-        self.update_views()
+        return self.fta_app.move_top_event_down(self)
 
     def get_top_level_nodes(self):
-        """Return a list of all nodes that have no parent."""
-        all_nodes = self.get_all_nodes()
-        top_level = [node for node in all_nodes if not node.parents]
-        return top_level
+        return self.fta_app.get_top_level_nodes(self)
         
     def find_node_by_id_all(self, unique_id):
         for top in self.top_events:
@@ -2582,671 +1536,96 @@ class AutoMLApp:
         return None
 
     def get_hazop_by_name(self, name):
-        for d in self.hazop_docs:
-            if d.name == name:
-                return d
-        return None
+        return self.risk_app.get_hazop_by_name(self, name)
 
     def get_hara_by_name(self, name):
-        for d in self.hara_docs:
-            if d.name == name:
-                return d
-        return None
+        return self.risk_app.get_hara_by_name(self, name)
 
     def update_hara_statuses(self):
-        """Update each risk assessment document's status based on linked reviews."""
-        for doc in self.hara_docs:
-            status = "draft"
-            for review in self.reviews:
-                if doc.name in getattr(review, "hara_names", []):
-                    if review.mode == "joint" and review.approved and self.review_is_closed_for(review):
-                        status = "closed"
-                        break
-                    else:
-                        status = "in review"
-            doc.status = status
-            doc.approved = status == "closed"
+        return self.risk_app.update_hara_statuses(self)
 
     def update_fta_statuses(self):
-        """Update status for each top level event based on linked reviews."""
-        for te in self.top_events:
-            status = "draft"
-            for review in self.reviews:
-                if te.unique_id in getattr(review, "fta_ids", []):
-                    if review.mode == "joint" and review.approved and self.review_is_closed_for(review):
-                        status = "closed"
-                        break
-                    else:
-                        status = "in review"
-            te.status = status
+        return self.risk_app.update_fta_statuses(self)
 
     def get_safety_goal_asil(self, sg_name):
-        """Return the highest ASIL level for a safety goal name across approved risk assessments."""
-        best = "QM"
-        for doc in getattr(self, "hara_docs", []):
-            if not getattr(doc, "approved", False) and getattr(doc, "status", "") != "closed":
-                continue
-            for e in doc.entries:
-                if sg_name and sg_name == e.safety_goal and ASIL_ORDER.get(e.asil, 0) > ASIL_ORDER.get(best, 0):
-                    best = e.asil
-        for te in self.top_events:
-            if sg_name and (sg_name == te.user_name or sg_name == te.safety_goal_description):
-                if ASIL_ORDER.get(te.safety_goal_asil or "QM", 0) > ASIL_ORDER.get(best, 0):
-                    best = te.safety_goal_asil or "QM"
-        return best
+        return self.risk_app.get_safety_goal_asil(self, sg_name)
 
     def get_hara_goal_asil(self, sg_name):
-        """Return highest ASIL from all risk assessment entries for the given safety goal."""
-        best = "QM"
-        for doc in getattr(self, "hara_docs", []):
-            for e in doc.entries:
-                if sg_name and sg_name == e.safety_goal and ASIL_ORDER.get(e.asil, 0) > ASIL_ORDER.get(best, 0):
-                    best = e.asil
-        return best
+        return self.risk_app.get_hara_goal_asil(self, sg_name)
 
     def get_cyber_goal_cal(self, goal_id):
-        """Return highest CAL from risk assessments for the given cybersecurity goal."""
-        order = {level: idx for idx, level in enumerate(CAL_LEVEL_OPTIONS, start=1)}
-        best = CAL_LEVEL_OPTIONS[0]
-        for doc in getattr(self, "hara_docs", []):
-            for e in getattr(doc, "entries", []):
-                cyber = getattr(e, "cyber", None)
-                if not cyber or not cyber.cybersecurity_goal:
-                    continue
-                if goal_id and goal_id == cyber.cybersecurity_goal:
-                    cal = getattr(cyber, "cal", CAL_LEVEL_OPTIONS[0])
-                    if order.get(cal, 0) > order.get(best, 0):
-                        best = cal
-        return best
+        return self.risk_app.get_cyber_goal_cal(self, goal_id)
 
     def get_top_event_safety_goals(self, node):
-        """Return names of safety goals for top events containing ``node``."""
-        result = []
-        target = self.get_failure_mode_node(node)
-        for te in self.top_events:
-            if any(n.unique_id == target.unique_id for n in self.get_all_nodes(te)):
-                sg = te.safety_goal_description or te.user_name or ""
-                if sg:
-                    result.append(sg)
-        return result
+        return self.risk_app.get_top_event_safety_goals(self, node)
 
     def get_safety_goals_for_malfunctions(self, malfunctions: list[str]) -> list[str]:
-        """Return safety goal names for given malfunctions."""
-        goals = []
-        for te in self.top_events:
-            mal = getattr(te, "malfunction", "")
-            if mal and mal in malfunctions:
-                sg = te.safety_goal_description or te.user_name or ""
-                if sg and sg not in goals:
-                    goals.append(sg)
-        return goals
+        return self.risk_app.get_safety_goals_for_malfunctions(self, malfunctions)
 
     def is_malfunction_used(self, name: str) -> bool:
-        """Return True if the malfunction is used in any FTA or analysis."""
-        if not name:
-            return False
-        for te in self.top_events:
-            if getattr(te, "malfunction", "") == name:
-                return True
-        for n in self.get_all_nodes_in_model():
-            mals = [m.strip() for m in getattr(n, "fmeda_malfunction", "").split(";") if m.strip()]
-            if name in mals:
-                return True
-        return False
+        return self.risk_app.is_malfunction_used(self, name)
 
     def add_malfunction(self, name: str) -> None:
-        """Add a malfunction to the list if it does not already exist."""
-        self.push_undo_state()
-        if not name:
-            return
-        name = name.strip()
-        if not name:
-            return
-        exists = any(m.lower() == name.lower() for m in self.malfunctions)
-        append_unique_insensitive(self.malfunctions, name)
-        if not exists and not any(
-            getattr(te, "malfunction", "") == name for te in self.top_events
-        ):
-            # If there's exactly one top event with no malfunction yet,
-            # reuse it instead of creating a new node.
-            if len(self.top_events) == 1 and not getattr(self.top_events[0], "malfunction", ""):
-                self.top_events[0].malfunction = name
-                self.root_node = self.top_events[0]
-                self.update_views()
-            else:
-                self.create_top_event_for_malfunction(name)
+        return self.risk_app.add_malfunction(self, name)
 
     def add_fault(self, name: str) -> None:
-        """Add a fault to the list if not already present."""
-        self.push_undo_state()
-        append_unique_insensitive(self.faults, name)
+        return self.risk_app.add_fault(self, name)
 
     def add_failure(self, name: str) -> None:
-        """Add a failure to the list if not already present."""
-        self.push_undo_state()
-        append_unique_insensitive(self.failures, name)
+        return self.risk_app.add_failure(self, name)
 
     def add_hazard(self, name: str, severity: int | str = 1) -> None:
-        """Add a hazard to the list if not already present."""
-        self.push_undo_state()
-        append_unique_insensitive(self.hazards, name)
-        if isinstance(severity, str):
-            try:
-                severity = int(severity)
-            except Exception:
-                severity = 1
-        if name not in self.hazard_severity:
-            self.hazard_severity[name] = int(severity)
+        return self.risk_app.add_hazard(self, name, severity)
 
-    def add_triggering_condition(self, name: str) -> None:
-        """Add a triggering condition to the repository."""
-        self.push_undo_state()
-        append_unique_insensitive(self.triggering_conditions, name or "")
-        self.update_views()
-
-    def add_functional_insufficiency(self, name: str) -> None:
-        """Add a functional insufficiency to the repository."""
-        self.push_undo_state()
-        append_unique_insensitive(self.functional_insufficiencies, name or "")
-        self.update_views()
-
-    def delete_triggering_condition(self, name: str) -> None:
-        """Remove a triggering condition and update references."""
-        self.push_undo_state()
-        self.triggering_conditions = [tc for tc in self.triggering_conditions if tc != name]
-        for doc in self.fi2tc_docs + self.tc2fi_docs:
-            for e in doc.entries:
-                val = e.get("triggering_conditions", "")
-                new_val = self._remove_name_from_list(val, name)
-                if new_val != val:
-                    e["triggering_conditions"] = new_val
-        self.update_views()
-
-    def delete_functional_insufficiency(self, name: str) -> None:
-        """Remove a functional insufficiency and update references."""
-        self.push_undo_state()
-        self.functional_insufficiencies = [fi for fi in self.functional_insufficiencies if fi != name]
-        for doc in self.fi2tc_docs + self.tc2fi_docs:
-            for e in doc.entries:
-                val = e.get("functional_insufficiencies", "")
-                new_val = self._remove_name_from_list(val, name)
-                if new_val != val:
-                    e["functional_insufficiencies"] = new_val
-        self.update_views()
-
-    # --------------------------------------------------------------
-    # Rename helpers propagate changes across the entire model
-    # --------------------------------------------------------------
-    def _replace_in_mal_list(self, obj, old, new):
-        val = getattr(obj, "fmeda_malfunction", "")
-        if not val:
-            return
-        parts = []
-        changed = False
-        for m in val.split(";"):
-            m = m.strip()
-            if not m:
-                continue
-            if m == old:
-                parts.append(new)
-                changed = True
-            else:
-                parts.append(m)
-        if changed:
-            obj.fmeda_malfunction = ";".join(parts)
-
-    def _replace_entry_mal(self, entry, old, new):
-        val = getattr(entry, "fmeda_malfunction", "")
-        if val:
-            parts = [new if m.strip() == old else m.strip() for m in val.split(";") if m.strip()]
-            if ";".join(parts) != val:
-                entry.fmeda_malfunction = ";".join(parts)
 
     def rename_malfunction(self, old: str, new: str) -> None:
-        """Rename a malfunction and update all references."""
-        self.push_undo_state()
-        if not old or old == new:
-            return
-        for i, m in enumerate(self.malfunctions):
-            if m == old:
-                self.malfunctions[i] = new
-        for te in self.top_events + getattr(self, "cta_events", []) + getattr(self, "paa_events", []):
-            if getattr(te, "malfunction", "") == old:
-                te.malfunction = new
-        for n in self.get_all_nodes_in_model():
-            self._replace_in_mal_list(n, old, new)
-        for doc in self.hazop_docs:
-            for e in doc.entries:
-                if getattr(e, "malfunction", "") == old:
-                    e.malfunction = new
-        for d in self.fmeas:
-            for e in d.get("entries", []):
-                self._replace_entry_mal(e, old, new)
-        for d in self.fmedas:
-            for e in d.get("entries", []):
-                self._replace_entry_mal(e, old, new)
-        self.update_views()
-        self._update_shared_product_goals()
+        return self.risk_app.rename_malfunction(self, old, new)
 
     def _update_shared_product_goals(self):
-        groups = {}
-        for te in self.top_events + getattr(self, "cta_events", []) + getattr(self, "paa_events", []):
-            mal = getattr(te, "malfunction", "")
-            if mal:
-                groups.setdefault(mal, []).append(te)
-        self.shared_product_goals = getattr(self, "shared_product_goals", {})
-        for mal, events in groups.items():
-            if len(events) > 1:
-                pg = self.shared_product_goals.get(mal)
-                if not pg:
-                    pg = {"name": events[0].user_name}
-                    self.shared_product_goals[mal] = pg
-                for e in events:
-                    e.user_name = pg["name"]
-                    e.name_readonly = True
-                    e.product_goal = pg
-            else:
-                self.shared_product_goals.pop(mal, None)
-                ev = events[0]
-                ev.name_readonly = False
-                ev.product_goal = None
+        return self.risk_app._update_shared_product_goals(self)
 
     def rename_hazard(self, old: str, new: str) -> None:
-        self.push_undo_state()
-        if not old or old == new:
-            return
-        for i, h in enumerate(self.hazards):
-            if h == old:
-                self.hazards[i] = new
-        if old in self.hazard_severity:
-            self.hazard_severity[new] = self.hazard_severity.pop(old)
-        for doc in self.hazop_docs:
-            for e in doc.entries:
-                if getattr(e, "hazard", "") == old:
-                    e.hazard = new
-        for doc in self.hara_docs:
-            for e in doc.entries:
-                if getattr(e, "hazard", "") == old:
-                    e.hazard = new
-        for doc in self.fi2tc_docs + self.tc2fi_docs:
-            for e in doc.entries:
-                if e.get("vehicle_effect", "") == old:
-                    e["vehicle_effect"] = new
-        self.update_views()
+        return self.risk_app.rename_hazard(self, old, new)
 
     def update_hazard_severity(self, hazard: str, severity: int | str) -> None:
-        self.push_undo_state()
-        try:
-            severity = int(severity)
-        except Exception:
-            severity = 1
-        self.hazard_severity[hazard] = severity
-        for doc in self.hara_docs:
-            for e in doc.entries:
-                if getattr(e, "hazard", "") == hazard:
-                    e.severity = severity
-        for doc in self.fi2tc_docs + self.tc2fi_docs:
-            for e in doc.entries:
-                if e.get("vehicle_effect", "") == hazard:
-                    e["severity"] = str(severity)
-        self.update_views()
+        return self.risk_app.update_hazard_severity(self, hazard, severity)
 
     def rename_fault(self, old: str, new: str) -> None:
-        self.push_undo_state()
-        if not old or old == new:
-            return
-        for i, f in enumerate(self.faults):
-            if f == old:
-                self.faults[i] = new
-        for n in self.get_all_nodes_in_model():
-            if getattr(n, "fault_ref", "") == old:
-                n.fault_ref = new
-        for be in self.get_all_fmea_entries():
-            causes = [c.strip() for c in getattr(be, "fmea_cause", "").split(";")]
-            changed = False
-            for idx, c in enumerate(causes):
-                if c == old:
-                    causes[idx] = new
-                    changed = True
-            if changed:
-                be.fmea_cause = ";".join([c for c in causes if c])
-        self.update_views()
+        return self.risk_app.rename_fault(self, old, new)
 
     def rename_failure(self, old: str, new: str) -> None:
-        self.push_undo_state()
-        if not old or old == new:
-            return
-        for i, fl in enumerate(self.failures):
-            if fl == old:
-                self.failures[i] = new
-        for be in self.get_all_fmea_entries():
-            if getattr(be, "fmea_effect", "") == old:
-                be.fmea_effect = new
-        for n in self.get_all_nodes_in_model():
-            if getattr(n, "fmea_effect", "") == old:
-                n.fmea_effect = new
-        self.update_views()
+        return self.risk_app.rename_failure(self, old, new)
 
-    def _replace_name_in_list(self, value: str, old: str, new: str) -> str:
-        parts = []
-        changed = False
-        for p in value.split(";"):
-            p = p.strip()
-            if not p:
-                continue
-            if p == old:
-                parts.append(new)
-                changed = True
-            else:
-                parts.append(p)
-        return ";".join(parts) if changed else value
-
-    def _remove_name_from_list(self, value: str, name: str) -> str:
-        parts = []
-        for p in value.split(";"):
-            p = p.strip()
-            if p and p != name:
-                parts.append(p)
-        return ";".join(parts)
 
     def add_triggering_condition(self, name: str) -> None:
-        self.push_undo_state()
-        name = (name or "").strip()
-        if not name or name in self.triggering_conditions:
-            return
-        node = FaultTreeNode(name, "Triggering Condition")
-        self.triggering_condition_nodes.append(node)
-        if name not in self.triggering_conditions:
-            self.triggering_conditions.append(name)
-        self.update_triggering_condition_list()
-        self.update_views()
+        return self.risk_app.add_triggering_condition(self, name)
 
     def delete_triggering_condition(self, name: str) -> None:
-        self.push_undo_state()
-        self.triggering_condition_nodes = [
-            n for n in self.triggering_condition_nodes if n.user_name != name
-        ]
-        for doc in self.fi2tc_docs + self.tc2fi_docs:
-            for e in doc.entries:
-                val = e.get("triggering_conditions", "")
-                new_val = self._remove_name_from_list(val, name)
-                if new_val != val:
-                    e["triggering_conditions"] = new_val
-        if name in self.triggering_conditions:
-            self.triggering_conditions.remove(name)
-        self.update_triggering_condition_list()
-        self.update_views()
+        return self.risk_app.delete_triggering_condition(self, name)
 
     def rename_triggering_condition(self, old: str, new: str) -> None:
-        self.push_undo_state()
-        if not old or old == new:
-            return
-        for n in self.get_all_triggering_conditions():
-            if n.user_name == old:
-                n.user_name = new
-        for doc in self.fi2tc_docs + self.tc2fi_docs:
-            for e in doc.entries:
-                val = e.get("triggering_conditions", "")
-                new_val = self._replace_name_in_list(val, old, new)
-                if new_val != val:
-                    e["triggering_conditions"] = new_val
-        if old in self.triggering_conditions:
-            idx = self.triggering_conditions.index(old)
-            self.triggering_conditions[idx] = new
-        self.update_triggering_condition_list()
-        self.update_views()
+        return self.risk_app.rename_triggering_condition(self, old, new)
 
     def add_functional_insufficiency(self, name: str) -> None:
-        self.push_undo_state()
-        name = (name or "").strip()
-        if not name or name in self.functional_insufficiencies:
-            return
-        node = FaultTreeNode(name, "Functional Insufficiency")
-        node.gate_type = "AND"
-        self.functional_insufficiency_nodes.append(node)
-        if name not in self.functional_insufficiencies:
-            self.functional_insufficiencies.append(name)
-        self.update_functional_insufficiency_list()
-        self.update_views()
+        return self.risk_app.add_functional_insufficiency(self, name)
 
     def delete_functional_insufficiency(self, name: str) -> None:
-        self.push_undo_state()
-        self.functional_insufficiency_nodes = [
-            n for n in self.functional_insufficiency_nodes if n.user_name != name
-        ]
-        for doc in self.fi2tc_docs + self.tc2fi_docs:
-            for e in doc.entries:
-                val = e.get("functional_insufficiencies", "")
-                new_val = self._remove_name_from_list(val, name)
-                if new_val != val:
-                    e["functional_insufficiencies"] = new_val
-        if name in self.functional_insufficiencies:
-            self.functional_insufficiencies.remove(name)
-        self.update_functional_insufficiency_list()
-        self.update_views()
+        return self.risk_app.delete_functional_insufficiency(self, name)
 
     def rename_functional_insufficiency(self, old: str, new: str) -> None:
-        self.push_undo_state()
-        if not old or old == new:
-            return
-        for n in self.get_all_functional_insufficiencies():
-            if n.user_name == old:
-                n.user_name = new
-        for doc in self.fi2tc_docs + self.tc2fi_docs:
-            for e in doc.entries:
-                val = e.get("functional_insufficiencies", "")
-                new_val = self._replace_name_in_list(val, old, new)
-                if new_val != val:
-                    e["functional_insufficiencies"] = new_val
-        if old in self.functional_insufficiencies:
-            idx = self.functional_insufficiencies.index(old)
-            self.functional_insufficiencies[idx] = new
-        self.update_functional_insufficiency_list()
-        self.update_views()
+        return self.risk_app.rename_functional_insufficiency(self, old, new)
 
     def calculate_fmeda_metrics(self, events):
-        """Return ASIL and FMEDA metrics for the given events."""
-        total = 0.0
-        unc_spf = 0.0
-        unc_lpf = 0.0
-        asil = "QM"
-        for be in events:
-            src = self.get_failure_mode_node(be)
-            fit_mode = getattr(be, "fmeda_fit", 0.0)
-            total += fit_mode
-            if src.fmeda_fault_type == "permanent":
-                unc_spf += fit_mode * (1 - src.fmeda_diag_cov)
-            else:
-                unc_lpf += fit_mode * (1 - src.fmeda_diag_cov)
-            sg = getattr(src, "fmeda_safety_goal", "")
-            sgs = self.get_top_event_safety_goals(src)
-            if sgs:
-                sg = ", ".join(sgs)
-            a = self.get_safety_goal_asil(sg)
-            if ASIL_ORDER.get(a, 0) > ASIL_ORDER.get(asil, 0):
-                asil = a
-        dc = (total - (unc_spf + unc_lpf)) / total if total else 0.0
-        self.reliability_total_fit = total
-        self.reliability_dc = dc
-        self.spfm = unc_spf
-        self.lpfm = unc_lpf
-        spfm_metric = 1 - unc_spf / total if total else 0.0
-        lpfm_metric = 1 - unc_lpf / total if total else 0.0
-        return asil, dc, spfm_metric, lpfm_metric
+        return self.risk_app.calculate_fmeda_metrics(self, events)
 
     def compute_fmeda_metrics(self, events):
-        """Return aggregate and per-goal FMEDA metrics."""
-        comp_fit = component_fit_map(self.reliability_components)
-        goal_metrics = {}
-        total = 0.0
-        spf_total = 0.0
-        lpf_total = 0.0
-        asil = "QM"
-        for be in events:
-            src = self.get_failure_mode_node(be)
-            goals = self.get_top_event_safety_goals(src) or [getattr(src, "fmeda_safety_goal", "")]
-            comp_name = self.get_component_name_for_node(src)
-            fit = comp_fit.get(comp_name)
-            frac = getattr(src, "fmeda_fault_fraction", 0.0)
-            if frac > 1.0:
-                frac /= 100.0
-            value = fit * frac if fit is not None else getattr(src, "fmeda_fit", 0.0)
-            fault_spf = value * (1 - src.fmeda_diag_cov) if src.fmeda_fault_type == "permanent" else 0.0
-            fault_lpf = value * (1 - src.fmeda_diag_cov) if src.fmeda_fault_type != "permanent" else 0.0
-            for sg in goals:
-                gm = goal_metrics.setdefault(
-                    sg,
-                    {
-                        "total": 0.0,
-                        "spfm_raw": 0.0,
-                        "lpfm_raw": 0.0,
-                        "asil": self.get_safety_goal_asil(sg),
-                    },
-                )
-                gm["total"] += value
-                gm["spfm_raw"] += fault_spf
-                gm["lpfm_raw"] += fault_lpf
-            total += value
-            spf_total += fault_spf
-            lpf_total += fault_lpf
-            for sg in goals:
-                a = self.get_safety_goal_asil(sg)
-                if ASIL_ORDER.get(a, 0) > ASIL_ORDER.get(asil, 0):
-                    asil = a
-
-        for sg, vals in goal_metrics.items():
-            t = vals["total"]
-            spf = vals["spfm_raw"]
-            lpf = vals["lpfm_raw"]
-            dc = (t - (spf + lpf)) / t if t else 0.0
-            spfm_metric = 1 - spf / t if t else 0.0
-            lpfm_metric = 1 - lpf / t if t else 0.0
-            thresh = ASIL_TARGETS.get(vals["asil"], ASIL_TARGETS["QM"])
-            vals.update(
-                {
-                    "dc": dc,
-                    "spfm_metric": spfm_metric,
-                    "lpfm_metric": lpfm_metric,
-                    "ok_dc": dc >= thresh["dc"],
-                    "ok_spfm": spfm_metric >= thresh["spfm"],
-                    "ok_lpfm": lpfm_metric >= thresh["lpfm"],
-                }
-            )
-
-        dc_total = (total - (spf_total + lpf_total)) / total if total else 0.0
-        spfm_metric_total = 1 - spf_total / total if total else 0.0
-        lpfm_metric_total = 1 - lpf_total / total if total else 0.0
-        thresh_total = ASIL_TARGETS.get(asil, ASIL_TARGETS["QM"])
-
-        self.reliability_total_fit = total
-        self.reliability_dc = dc_total
-        self.spfm = spf_total
-        self.lpfm = lpf_total
-
-        return {
-            "total": total,
-            "spfm_raw": spf_total,
-            "lpfm_raw": lpf_total,
-            "dc": dc_total,
-            "spfm_metric": spfm_metric_total,
-            "lpfm_metric": lpfm_metric_total,
-            "asil": asil,
-            "ok_dc": dc_total >= thresh_total["dc"],
-            "ok_spfm": spfm_metric_total >= thresh_total["spfm"],
-            "ok_lpfm": lpfm_metric_total >= thresh_total["lpfm"],
-            "goal_metrics": goal_metrics,
-        }
+        return self.risk_app.compute_fmeda_metrics(self, events)
 
     def sync_hara_to_safety_goals(self):
-        """Propagate risk assessment values to top events, inheriting ASILs from assessment rows."""
-        sg_data = {}
-        sg_asil = {}
-        toolbox = getattr(self, "safety_toolbox", None)
-        for doc in getattr(self, "hara_docs", []):
-            approved = getattr(doc, "approved", False) or getattr(doc, "status", "") == "closed"
-            for e in doc.entries:
-                mal = getattr(e, "malfunction", "")
-                if not mal:
-                    continue
-                data = sg_data.setdefault(
-                    mal,
-                    {"asil": "QM", "severity": 1, "cont": 1, "exp": 1, "sg": "", "approved": False},
-                )
-                if ASIL_ORDER.get(e.asil, 0) > ASIL_ORDER.get(data["asil"], 0):
-                    data["asil"] = e.asil
-                    data["sg"] = e.safety_goal
-                if e.severity > data["severity"]:
-                    data["severity"] = e.severity
-                if e.controllability > data["cont"]:
-                    data["cont"] = e.controllability
-                if e.exposure > data["exp"]:
-                    data["exp"] = e.exposure
-                if approved:
-                    data["approved"] = True
-                if e.safety_goal and (
-                    not toolbox
-                    or toolbox.can_propagate(
-                        "Risk Assessment",
-                        "Product Goal Specification",
-                        reviewed=approved,
-                        joint_review=approved,
-                    )
-                ):
-                    best = sg_asil.get(e.safety_goal, "QM")
-                    if ASIL_ORDER.get(e.asil, 0) > ASIL_ORDER.get(best, 0):
-                        sg_asil[e.safety_goal] = e.asil
-
-        for te in self.top_events:
-            mal = getattr(te, "malfunction", "")
-            data = sg_data.get(mal)
-            if data:
-                propagate = False
-                if (
-                    not toolbox
-                    or toolbox.can_propagate(
-                        "Risk Assessment",
-                        "FTA",
-                        reviewed=data.get("approved", False),
-                        joint_review=data.get("approved", False),
-                    )
-                ):
-                    if getattr(te, "status", "draft") != "closed":
-                        propagate = True
-                    elif data.get("approved"):
-                        propagate = True
-                        te.status = "draft"
-                        self.invalidate_reviews_for_fta(te.unique_id)
-                if propagate:
-                    te.safety_goal_description = data["sg"]
-                    te.severity = data["severity"]
-                    te.controllability = data["cont"]
-                    te.exposure = data["exp"]
-                    te.update_validation_target()
-            sg_name = te.safety_goal_description
-            asil = sg_asil.get(sg_name)
-            flag = data.get("approved", False) if data else False
-            if toolbox and not toolbox.can_propagate(
-                "FTA", "Product Goal Specification", reviewed=flag, joint_review=flag
-            ):
-                asil = None
-            if asil and ASIL_ORDER.get(asil, 0) > ASIL_ORDER.get(te.safety_goal_asil or "QM", 0):
-                te.safety_goal_asil = asil
+        return self.risk_app.sync_hara_to_safety_goals(self)
 
     def sync_cyber_risk_to_goals(self):
-        """Aggregate CAL values from risk assessments into cybersecurity goals."""
-        goal_map = {g.goal_id: g for g in getattr(self, "cybersecurity_goals", [])}
-        for g in goal_map.values():
-            g.risk_assessments = []
-        for doc in getattr(self, "hara_docs", []):
-            for e in getattr(doc, "entries", []):
-                cyber = getattr(e, "cyber", None)
-                if not cyber or not cyber.cybersecurity_goal:
-                    continue
-                cg = goal_map.get(cyber.cybersecurity_goal)
-                if cg is not None:
-                    cg.risk_assessments.append({"name": doc.name, "cal": cyber.cal})
-        for g in goal_map.values():
-            g.compute_cal()
+        return self.risk_app.sync_cyber_risk_to_goals(self)
 
     def edit_selected(self):
         sel = self.analysis_tree.selection()
@@ -8222,186 +6601,20 @@ class AutoMLApp:
                                font=self.diagram_font)
 
     def save_diagram_png(self):
-        self.diagram_controller.save_diagram_png()
+        self.diagram_export_app.save_diagram_png()
 
     def on_treeview_click(self, event):
-        sel = self.analysis_tree.selection()
-        if not sel:
-            return
-        try:
-            node_id = int(self.analysis_tree.item(sel[0], "tags")[0])
-        except (IndexError, ValueError):
-            return
-        node = self.find_node_by_id_all(node_id)
-        if node:
-            self.open_page_diagram(node)
+        self.tree_app.on_treeview_click(self, event)
 
     def on_analysis_tree_double_click(self, event):
-        item = (
-            self.analysis_tree.identify_row(event.y)
-            if event is not None
-            else self.analysis_tree.focus()
-        )
-        if not item:
-            return
-        self.analysis_tree.focus(item)
-        tags = self.analysis_tree.item(item, "tags")
-        kind = tags[0] if tags else None
-        ident = tags[1] if tags and len(tags) > 1 else None
-        if kind in {"fmea", "fmeda", "hazop", "hara", "stpa", "threat", "fi2tc", "tc2fi", "jrev", "gov"} and ident is not None:
-            idx = int(ident)
-            if kind == "fmea":
-                self.show_fmea_table(self.fmeas[idx])
-            elif kind == "fmeda":
-                self.show_fmea_table(self.fmedas[idx], fmeda=True)
-            elif kind == "hazop":
-                self.open_hazop_window()
-                if hasattr(self, "_hazop_window"):
-                    doc = self.hazop_docs[idx]
-                    self._hazop_window.doc_var.set(doc.name)
-                    self._hazop_window.select_doc()
-            elif kind == "hara":
-                self.open_risk_assessment_window()
-                if hasattr(self, "_risk_window"):
-                    doc = self.hara_docs[idx]
-                    self._risk_window.doc_var.set(doc.name)
-                    self._risk_window.select_doc()
-            elif kind == "stpa":
-                self.open_stpa_window()
-                if hasattr(self, "_stpa_window"):
-                    doc = self.stpa_docs[idx]
-                    self._stpa_window.doc_var.set(doc.name)
-                    self._stpa_window.select_doc()
-            elif kind == "threat":
-                self.open_threat_window()
-                if hasattr(self, "_threat_window"):
-                    doc = self.threat_docs[idx]
-                    self._threat_window.doc_var.set(doc.name)
-                    self._threat_window.select_doc()
-            elif kind == "fi2tc":
-                self.open_fi2tc_window()
-                if hasattr(self, "_fi2tc_window"):
-                    doc = self.fi2tc_docs[idx]
-                    self._fi2tc_window.doc_var.set(doc.name)
-                    self._fi2tc_window.select_doc()
-            elif kind == "tc2fi":
-                self.open_tc2fi_window()
-                if hasattr(self, "_tc2fi_window"):
-                    doc = self.tc2fi_docs[idx]
-                    self._tc2fi_window.doc_var.set(doc.name)
-                    self._tc2fi_window.select_doc()
-            elif kind == "jrev":
-                if 0 <= idx < len(getattr(self, "joint_reviews", [])):
-                    review = self.joint_reviews[idx]
-                    self.review_data = review
-                    self.open_review_document(review)
-                    self.open_review_toolbox()
-            elif kind == "gov":
-                self.open_management_window(idx)
-        elif kind == "gsn" and ident is not None:
-            diag = getattr(self, "gsn_diagram_map", {}).get(ident)
-            if diag:
-                self.open_gsn_diagram(diag)
-        elif kind == "gsnmod":
-            self.manage_gsn()
-        elif kind == "reqs":
-            self.show_requirements_editor()
-        elif kind == "reqexp":
-            self.show_requirements_explorer()
-        elif kind == "sg":
-            self.show_product_goals_editor()
-        elif kind == "fta" and ident is not None:
-            te = next((t for t in self.top_events if t.unique_id == int(ident)), None)
-            if te:
-                self.diagram_mode = "FTA"
-                self.ensure_fta_tab()
-                self.doc_nb.select(self.canvas_tab)
-                self.open_page_diagram(te)
-        elif kind == "cta" and ident is not None:
-            te = next((t for t in getattr(self, "cta_events", []) if t.unique_id == int(ident)), None)
-            if te:
-                self.diagram_mode = "CTA"
-                self.ensure_fta_tab()
-                self.doc_nb.select(self.canvas_tab)
-                self.open_page_diagram(te)
-        elif kind == "paa" and ident is not None:
-            te = next((t for t in getattr(self, "paa_events", []) if t.unique_id == int(ident)), None)
-            if te:
-                self.diagram_mode = "PAA"
-                self.ensure_fta_tab()
-                self.doc_nb.select(self.canvas_tab)
-                self.open_page_diagram(te)
-        elif kind == "safetycase":
-            self.manage_safety_cases()
-        elif kind == "safetyconcept":
-            self.show_safety_concept_editor()
-        elif kind == "itemdef":
-            self.show_item_definition_editor()
-        elif kind == "arch":
-            self.open_arch_window(ident)
-        elif kind == "pkg":
-            self.manage_architecture()
-        else:
-            parent = item
-            while parent:
-                if (
-                    self.analysis_tree.item(parent, "text")
-                    == "Safety & Security Governance Diagrams"
-                ):
-                    self.manage_safety_management()
-                    return
-                parent = self.analysis_tree.parent(parent)
+        self.tree_app.on_analysis_tree_double_click(self, event)
 
     def on_analysis_tree_right_click(self, event):
-        iid = self.analysis_tree.identify_row(event.y)
-        if not iid:
-            return
-        self.analysis_tree.selection_set(iid)
-        self.analysis_tree.focus(iid)
-        menu = tk.Menu(self.analysis_tree, tearoff=0)
-        menu.add_command(label="Rename", command=self.rename_selected_tree_item)
-        menu.tk_popup(event.x_root, event.y_root)
+        self.tree_app.on_analysis_tree_right_click(self, event)
 
     def on_analysis_tree_select(self, _event):
         """Update property view when a tree item is selected."""
-        if not hasattr(self, "prop_view"):
-            return
-        item = self.analysis_tree.focus()
-        if not item:
-            return
-        tags = self.analysis_tree.item(item, "tags")
-        name = self.analysis_tree.item(item, "text")
-        meta = {"Name": name}
-        if tags:
-            meta["Type"] = tags[0]
-            if len(tags) > 1:
-                ident = tags[1]
-                meta["ID"] = ident
-                repo = SysMLRepository.get_instance()
-                elem = repo.elements.get(ident)
-                if elem:
-                    meta.update(
-                        {
-                            "Type": elem.elem_type,
-                            "Author": getattr(elem, "author", ""),
-                            "Created": getattr(elem, "created", ""),
-                            "Modified": getattr(elem, "modified", ""),
-                            "ModifiedBy": getattr(elem, "modified_by", ""),
-                        }
-                    )
-                else:
-                    diag = repo.diagrams.get(ident)
-                    if diag:
-                        meta.update(
-                            {
-                                "Type": diag.diag_type,
-                                "Author": getattr(diag, "author", ""),
-                                "Created": getattr(diag, "created", ""),
-                                "Modified": getattr(diag, "modified", ""),
-                                "ModifiedBy": getattr(diag, "modified_by", ""),
-                            }
-                        )
-        self.show_properties(meta=meta)
+        self.tree_app.on_analysis_tree_select(self, _event)
 
     def show_properties(self, obj=None, meta=None):
         """Display metadata for *obj* or *meta* dictionary in the properties tab."""
@@ -8457,125 +6670,7 @@ class AutoMLApp:
                         self.status_meta_vars[k].set(v)
 
     def rename_selected_tree_item(self):
-        item = self.analysis_tree.focus()
-        tags = self.analysis_tree.item(item, "tags")
-        if len(tags) != 2:
-            return
-        kind, ident = tags[0], tags[1]
-        repo = SysMLRepository.get_instance()
-        current = ""
-        node = None
-        if kind in {"fmea", "fmeda", "hazop", "hara", "fi2tc", "tc2fi", "jrev"}:
-            idx = int(ident)
-            if kind == "fmea":
-                current = self.fmeas[idx]["name"]
-            elif kind == "fmeda":
-                current = self.fmedas[idx]["name"]
-            elif kind == "hazop":
-                current = self.hazop_docs[idx].name
-            elif kind == "hara":
-                current = self.hara_docs[idx].name
-            elif kind == "fi2tc":
-                current = self.fi2tc_docs[idx].name
-            elif kind == "tc2fi":
-                current = self.tc2fi_docs[idx].name
-            elif kind == "jrev":
-                current = self.joint_reviews[idx].name
-        elif kind == "gsn":
-            diag = getattr(self, "gsn_diagram_map", {}).get(ident)
-            if not diag:
-                return
-            current = diag.root.user_name
-        elif kind == "gsnmod":
-            module = getattr(self, "gsn_module_map", {}).get(ident)
-            if not module:
-                return
-            current = module.name
-        elif kind == "arch":
-            diag = repo.diagrams.get(ident)
-            current = diag.name if diag else ""
-        elif kind == "gov":
-            idx = int(ident)
-            current = self.management_diagrams[idx].name
-        elif kind == "fta":
-            node = next((t for t in self.top_events if t.unique_id == int(ident)), None)
-            current = node.user_name if node else ""
-        elif kind == "pkg":
-            pkg = repo.elements.get(ident)
-            current = pkg.name if pkg else ""
-        else:
-            return
-        new = simpledialog.askstring("Rename", "Enter new name:", initialvalue=current)
-        if not new:
-            return
-        if kind == "fmea":
-            old = self.fmeas[idx]["name"]
-            self.fmeas[idx]["name"] = new
-            self.safety_mgmt_toolbox.rename_document("FMEA", old, new)
-        elif kind == "fmeda":
-            doc = self.fmedas[idx]
-            self.fmeda_manager.rename_fmeda(doc, new)
-        elif kind == "hazop":
-            old = self.hazop_docs[idx].name
-            self.hazop_docs[idx].name = new
-            self.safety_mgmt_toolbox.rename_document("HAZOP", old, new)
-        elif kind == "hara":
-            old = self.hara_docs[idx].name
-            self.hara_docs[idx].name = new
-            self.safety_mgmt_toolbox.rename_document("Risk Assessment", old, new)
-        elif kind == "fi2tc":
-            old = self.fi2tc_docs[idx].name
-            self.fi2tc_docs[idx].name = new
-            self.safety_mgmt_toolbox.rename_document("FI2TC", old, new)
-        elif kind == "tc2fi":
-            old = self.tc2fi_docs[idx].name
-            self.tc2fi_docs[idx].name = new
-            self.safety_mgmt_toolbox.rename_document("TC2FI", old, new)
-        elif kind == "fta":
-            node = next((t for t in self.top_events if t.unique_id == int(ident)), None)
-            if node:
-                old = node.user_name
-                node.user_name = new
-                analysis = (
-                    "Prototype Assurance Analysis"
-                    if getattr(self, "diagram_mode", "") == "PAA"
-                    else "FTA"
-                )
-                self.safety_mgmt_toolbox.rename_document(analysis, old, new)
-        elif kind == "arch" and repo.diagrams.get(ident):
-            repo.diagrams[ident].name = new
-        elif kind == "gov":
-            self.management_diagrams[idx].name = new
-        elif kind == "gsn":
-            diag = self.gsn_diagram_map.get(ident)
-            if diag:
-                diag.root.user_name = new
-        elif kind == "gsnmod":
-            module = self.gsn_module_map.get(ident)
-            if module:
-                module.name = new
-        elif kind == "jrev":
-            if any(r.name == new for r in self.reviews if r is not self.joint_reviews[idx]):
-                messagebox.showerror("Review", "Name already exists")
-                return
-            old = self.joint_reviews[idx].name
-            self.joint_reviews[idx].name = new
-            self.safety_mgmt_toolbox.rename_document("Joint Review", old, new)
-        elif kind == "fta" and node:
-            old = node.name
-            node.user_name = new
-            if hasattr(self, "safety_mgmt_toolbox"):
-                analysis = (
-                    "Prototype Assurance Analysis"
-                    if getattr(self, "diagram_mode", "") == "PAA"
-                    else "FTA"
-                )
-                self.safety_mgmt_toolbox.rename_document(analysis, old, node.name)
-        elif kind == "pkg" and repo.elements.get(ident):
-            repo.elements[ident].name = new
-        self.update_views()
-        if hasattr(self, "_arch_window") and self._arch_window.winfo_exists():
-            self._arch_window.populate()
+        self.tree_app.rename_selected_tree_item(self)
 
     def on_tool_list_double_click(self, event):
         lb = event.widget
@@ -11933,99 +10028,7 @@ class AutoMLApp:
         refresh()
 
     def show_hazard_list(self):
-        """Open a tab to manage the list of hazards."""
-        if hasattr(self, "_haz_tab") and self._haz_tab.winfo_exists():
-            self.doc_nb.select(self._haz_tab)
-            return
-        self._haz_tab = self._new_tab("Hazards")
-        win = self._haz_tab
-
-        # Load hazards from existing documents once when the tab is opened.
-        self.update_hazard_list()
-
-        tree = ttk.Treeview(win, columns=("Hazard", "Severity"), show="headings")
-        tree.heading("Hazard", text="Hazard")
-        tree.heading("Severity", text="Severity")
-        tree.column("Hazard", width=200)
-        tree.column("Severity", width=80)
-        tree.pack(side=tk.LEFT, fill=tk.BOTH, expand=True)
-
-        class _HazardDialog(simpledialog.Dialog):
-            """Prompt for a hazard name and severity."""
-
-            def __init__(self, parent, title: str, name: str = "", severity: str = "1"):
-                self._name = name
-                self._severity = severity
-                super().__init__(parent, title=title)
-
-            def body(self, master):
-                self.resizable(False, False)
-                ttk.Label(master, text="Name:").grid(row=0, column=0, sticky="e")
-                self.name_var = tk.StringVar(value=self._name)
-                name_entry = ttk.Entry(master, textvariable=self.name_var)
-                name_entry.grid(row=0, column=1, padx=5, pady=5)
-                ttk.Label(master, text="Severity:").grid(row=1, column=0, sticky="e")
-                self.sev_var = tk.StringVar(value=self._severity)
-                ttk.Combobox(
-                    master,
-                    textvariable=self.sev_var,
-                    values=["1", "2", "3"],
-                    state="readonly",
-                ).grid(row=1, column=1, padx=5, pady=5)
-                return name_entry
-
-            def apply(self):
-                self.result = (
-                    self.name_var.get().strip(),
-                    self.sev_var.get().strip(),
-                )
-
-        def refresh():
-            tree.delete(*tree.get_children())
-            for h in self.hazards:
-                tree.insert("", "end", values=(h, self.hazard_severity.get(h, "")))
-
-        def add():
-            dlg = _HazardDialog(win, "Add Hazard")
-            if not dlg.result:
-                return
-            name, sev = dlg.result
-            if name:
-                self.add_hazard(name, sev)
-                refresh()
-
-        def rename():
-            sel = tree.focus()
-            if not sel:
-                return
-            current, sev = tree.item(sel, "values")[:2]
-            dlg = _HazardDialog(win, "Edit Hazard", current, str(sev))
-            if not dlg.result:
-                return
-            name, sev_val = dlg.result
-            if name:
-                if name != current:
-                    self.rename_hazard(current, name)
-                self.update_hazard_severity(name, sev_val)
-                refresh()
-
-        def delete():
-            sel = tree.focus()
-            if not sel:
-                return
-            current = tree.item(sel, "values")[0]
-            if messagebox.askyesno("Delete", f"Delete '{current}'?"):
-                self.hazards.remove(current)
-                self.hazard_severity.pop(current, None)
-                refresh()
-
-        btn = ttk.Frame(win)
-        btn.pack(side=tk.RIGHT, fill=tk.Y)
-        ttk.Button(btn, text="Add", command=add).pack(fill=tk.X)
-        ttk.Button(btn, text="Edit", command=rename).pack(fill=tk.X)
-        ttk.Button(btn, text="Delete", command=delete).pack(fill=tk.X)
-
-        refresh()
+        self.risk_app.show_hazard_list(self)
 
     def show_malfunction_editor(self):
         """Open a tab to manage global malfunctions."""
@@ -12188,8 +10191,7 @@ class AutoMLApp:
     # ------------------------------------------------------------------
 
     def show_hazard_editor(self):
-        """Backward compatible alias for :meth:`show_hazard_list`."""
-        self.show_hazard_list()
+        self.risk_app.show_hazard_editor(self)
 
     def show_fault_editor(self):
         """Backward compatible alias for :meth:`show_fault_list`."""
@@ -16348,87 +14350,34 @@ class AutoMLApp:
         refresh_libs()
 
     def open_reliability_window(self):
-        if hasattr(self, "_rel_tab") and self._rel_tab.winfo_exists():
-            self.doc_nb.select(self._rel_tab)
-        else:
-            self._rel_tab = self._new_tab("Reliability")
-            self._rel_window = ReliabilityWindow(self._rel_tab, self)
-            self._rel_window.pack(fill=tk.BOTH, expand=True)
-        self.refresh_all()
+        self.reliability_app.open_reliability_window(self)
 
     def open_fmeda_window(self):
-        self.show_fmeda_list()
-        self.refresh_all()
+        self.reliability_app.open_fmeda_window(self)
 
     def open_hazop_window(self):
-        if hasattr(self, "_hazop_tab") and self._hazop_tab.winfo_exists():
-            self.doc_nb.select(self._hazop_tab)
-        else:
-            self._hazop_tab = self._new_tab("HAZOP")
-            self._hazop_window = HazopWindow(self._hazop_tab, self)
-        self.refresh_all()
+        self.risk_app.open_hazop_window(self)
 
     def open_risk_assessment_window(self):
-        if hasattr(self, "_risk_tab") and self._risk_tab.winfo_exists():
-            self.doc_nb.select(self._risk_tab)
-        else:
-            self._risk_tab = self._new_tab("Risk Assessment")
-            self._risk_window = RiskAssessmentWindow(self._risk_tab, self)
-        self.refresh_all()
+        self.risk_app.open_risk_assessment_window(self)
 
     def open_stpa_window(self):
-        if hasattr(self, "_stpa_tab") and self._stpa_tab.winfo_exists():
-            self.doc_nb.select(self._stpa_tab)
-        else:
-            self._stpa_tab = self._new_tab("STPA")
-            self._stpa_window = StpaWindow(self._stpa_tab, self)
-        self.refresh_all()
+        self.risk_app.open_stpa_window(self)
 
     def open_threat_window(self):
-        if hasattr(self, "_threat_tab") and self._threat_tab.winfo_exists():
-            self.doc_nb.select(self._threat_tab)
-        else:
-            self._threat_tab = self._new_tab("Threat")
-            self._threat_window = ThreatWindow(self._threat_tab, self)
-        self.refresh_all()
+        self.risk_app.open_threat_window(self)
 
     def open_causal_bayesian_network_window(self):
-        """Open the Causal Bayesian Network analysis window."""
-        if hasattr(self, "_cbn_tab") and self._cbn_tab.winfo_exists():
-            self.doc_nb.select(self._cbn_tab)
-        else:
-            self._cbn_tab = self._new_tab("Causal Bayesian Network")
-            from gui.causal_bayesian_network_window import (
-                CausalBayesianNetworkWindow,
-            )
-
-            self._cbn_window = CausalBayesianNetworkWindow(self._cbn_tab, self)
-        self.refresh_all()
+        self.risk_app.open_causal_bayesian_network_window(self)
 
     def open_fi2tc_window(self):
-        if hasattr(self, "_fi2tc_tab") and self._fi2tc_tab.winfo_exists():
-            self.doc_nb.select(self._fi2tc_tab)
-        else:
-            self._fi2tc_tab = self._new_tab("FI2TC")
-            self._fi2tc_window = FI2TCWindow(self._fi2tc_tab, self)
-        self.refresh_all()
+        self.risk_app.open_fi2tc_window(self)
 
     def open_tc2fi_window(self):
-        if hasattr(self, "_tc2fi_tab") and self._tc2fi_tab.winfo_exists():
-            self.doc_nb.select(self._tc2fi_tab)
-        else:
-            self._tc2fi_tab = self._new_tab("TC2FI")
-            self._tc2fi_window = TC2FIWindow(self._tc2fi_tab, self)
-        self.refresh_all()
+        self.risk_app.open_tc2fi_window(self)
 
     def open_fault_prioritization_window(self):
-        if hasattr(self, "_fault_prio_tab") and self._fault_prio_tab.winfo_exists():
-            self.doc_nb.select(self._fault_prio_tab)
-        else:
-            self._fault_prio_tab = self._new_tab("Fault Prioritization")
-            from gui.fault_prioritization import FaultPrioritizationWindow
-            self._fault_prio_window = FaultPrioritizationWindow(self._fault_prio_tab, self)
-        self.refresh_all()
+        self.reliability_app.open_fault_prioritization_window(self)
 
     def open_safety_management_toolbox(self, show_diagrams: bool = True):
         """Open the Safety & Security Management editor and browser."""
@@ -16620,12 +14569,7 @@ class AutoMLApp:
                     child.redraw()
 
     def show_hazard_explorer(self):
-        if hasattr(self, "_haz_exp_tab") and self._haz_exp_tab.winfo_exists():
-            self.doc_nb.select(self._haz_exp_tab)
-        else:
-            self._haz_exp_tab = self._new_tab("Hazard Explorer")
-            self._haz_exp_window = HazardExplorerWindow(self._haz_exp_tab, self)
-            self._haz_exp_window.pack(fill=tk.BOTH, expand=True)
+        self.risk_app.show_hazard_explorer(self)
 
     def show_requirements_explorer(self):
         if hasattr(self, "_req_exp_tab") and self._req_exp_tab.winfo_exists():
@@ -17096,19 +15040,19 @@ class AutoMLApp:
         return create_icon(shape, color)
 
     def open_use_case_diagram(self):
-        self.diagram_controller.open_use_case_diagram()
+        self.use_case_diagram_app.open()
 
     def open_activity_diagram(self):
-        self.diagram_controller.open_activity_diagram()
+        self.activity_diagram_app.open()
 
     def open_block_diagram(self):
-        self.diagram_controller.open_block_diagram()
+        self.block_diagram_app.open()
 
     def open_internal_block_diagram(self):
-        self.diagram_controller.open_internal_block_diagram()
+        self.internal_block_diagram_app.open()
 
     def open_control_flow_diagram(self):
-        self.diagram_controller.open_control_flow_diagram()
+        self.control_flow_diagram_app.open()
 
     def manage_architecture(self):
         if hasattr(self, "_arch_tab") and self._arch_tab.winfo_exists():
